WIP split iplocallab.cc
