+++ rtengine/iplocallab.cc
@@ -2264,89 +2264,11 @@ void  ImProcFunctions::sigmoid_QJ(float Q, float &Qout, float middle_grey_contra
     Qout = value;
 }
 
-// Copyright 2018 Alberto Griggio <alberto.griggio@gmail.com>
+// find_gray moved to locallab/iplocallab_ciecam.cc
 
-float find_gray(float source_gray, float target_gray)
-{
-    // find a base such that log2lin(base, source_gray) = target_gray
-    // log2lin is (base^source_gray - 1) / (base - 1), so we solve
-    //
-    //  (base^source_gray - 1) / (base - 1) = target_gray, that is
-    //
-    //  base^source_gray - 1 - base * target_gray + target_gray = 0
-    //
-    // use a bisection method (maybe later change to Netwon)
+// mean_sig, log_encode, getAutoLogloc moved to locallab/iplocallab_log.cc
 
-    if (source_gray <= 0.f) {
-        return 0.f;
-    }
-
-    const auto f =
-    [ = ](float x) -> float {
-        return std::pow(x, source_gray) - 1.f - target_gray * x + target_gray;
-    };
-
-    // first find the interval we are interested in
-
-    float lo = 1.f;
-
-    while (f(lo) <= 0.f) {
-        lo *= 2.f;
-    }
-
-    float hi = lo * 2.f;
-
-    while (f(hi) >= 0.f) {
-        hi *= 2.f;
-    }
-
-    if (std::isinf(hi)) {
-        return 0.f;
-    }
-
-    // now search for a zero
-    for (int iter = 0; iter < 100; ++iter) {
-        float mid = lo + (hi - lo) / 2.f;
-        float v = f(mid);
-
-        if (std::abs(v) < 1e-4f || (hi - lo) / lo <= 1e-4f) {
-            return mid;
-        }
-
-        if (v > 0.f) {
-            lo = mid;
-        } else {
-            hi = mid;
-        }
-    }
-
-    return 0.f; // not found
-}
-
-void ImProcFunctions::mean_sig(const float* const * const savenormL, float &meanf, float &stdf, int xStart, int xEnd, int yStart, int yEnd) const
-{
-    const int size = (yEnd - yStart) * (xEnd - xStart);
-    // use double precision for large accumulations
-    double meand = 0.0;
-    double stdd = 0.0;
-#ifdef _OPENMP
-    #pragma omp parallel for reduction(+:meand, stdd) if(multiThread)
-#endif
-
-    for (int y = yStart; y < yEnd; ++y) {
-        for (int x = xStart; x < xEnd; ++x) {
-            meand += static_cast<double>(savenormL[y][x]);
-            stdd += SQR(static_cast<double>(savenormL[y][x]));
-        }
-    }
-
-    meand /= size;
-    stdd /= size;
-    stdd -= SQR(meand);
-    stdf = std::sqrt(stdd);
-    meanf = meand;
-}
-
+// Helper functions used by log encoding and other tools
 // taken from darktable
 inline float power_norm(float r, float g, float b)
 {
@@ -2357,7 +2279,7 @@ inline float power_norm(float r, float g, float b)
     float r2 = SQR(r);
     float g2 = SQR(g);
     float b2 = SQR(b);
-  
+
     float d = r2 + g2 + b2;
     float n = r * r2 + g * g2 + b * b2;
 
@@ -2386,409 +2308,20 @@ inline float norm2(float r, float g, float b, TMatrix ws)
 inline float norm_3(float r, float g, float b, TMatrix ws, float raplim)//lowers the equivalent luminance if the white point is high
 {
     constexpr float hi = std::numeric_limits<float>::max() / 100.f;
-    float pwn = 0.5f;//standard repartition between XYZ luminance and Out of gamut values 
+    float pwn = 0.5f;//standard repartition between XYZ luminance and Out of gamut values
     if (raplim < 1.2f) {//raplim : ratio between the normal value 'reasonable_limit_white_point' and reality
         pwn = 0.55f;//Tested on images with WP linear close to 4 - Near Sunset
     } else if (raplim < 1.5f) {//Very high White point
         pwn = 0.75f;//Tested on images with WP linear close to 5 or 6
     } else {
         pwn = 0.85f;//Tested on images with WP linear close to 6 and above //LEDs
-    }    
+    }
     return std::min(hi, (1.f - pwn) * power_norm(r, g, b) + pwn * Color::rgbLuminance(r, g, b, ws));//I reversed the action of the two components to better account for what happens out of gamut.
 }
 
 inline float norm(float r, float g, float b, TMatrix ws)
 {
     return (Color::rgbLuminance(r, g, b, ws));
-}
-
-
-// basic log encoding taken from ACESutil.Lin_to_Log2, from
-// https://github.com/ampas/aces-dev
-// (as seen on pixls.us)
-// copyright 2018 Alberto Griggio <alberto.griggio@gmail.com>
-void ImProcFunctions::log_encode(Imagefloat *rgb, struct local_params & lp, bool multiThread, int bfw, int bfh)
-{
-    // BENCHFUN
-        float gray = 0.1f;
-        float shadows_range = 0.f;
-        bool comprlog = 0.f;
-        float comprfactorlog = 0.f;
-        float dynamic_range = 1.f;
-        float targray = 0.1f;
-
-    bool satcontrol = false;
-        
-    if(lp.logena) {
-        gray = 0.01f * lp.sourcegray;
-        shadows_range = lp.blackev;
-        comprlog = lp.comprlo  > 0.f;
-        comprfactorlog = lp.comprlo;
-        dynamic_range = max(lp.whiteev - lp.blackev, 0.5f);
-        targray = lp.targetgray;
-        satcontrol = lp.satlog;
-
-    } else if (lp.cieena) {
-        gray = 0.01f * lp.sourcegraycie;
-        shadows_range = lp.blackevjz;
-        comprlog = lp.comprlocie  > 0.f;
-        comprfactorlog = lp.comprlocie;
-        dynamic_range = max(lp.whiteevjz - lp.blackevjz, 0.5f);
-        targray = lp.targetgraycie;
-        satcontrol = lp.satcie;
-    }
-    float comprthlog = 1.f;
-
-    const float noise = pow_F(2.f, -16.f);
-    const float log2 = xlogf(2.f);
-    const float base = targray > 1 && targray < 100 && dynamic_range > 0 ? find_gray(std::abs(shadows_range) / dynamic_range, 0.01f * targray) : 0.f;
-    const float linbase = rtengine::max(base, 2.f);//2 to avoid bad behavior
-    TMatrix ws = ICCStore::getInstance()->workingSpaceMatrix(params->icm.workingProfile);
-    float ac = -5.f;//max 4
-    float bc = 4.f;
-    if(comprlog < 0.6f) {
-        comprthlog = ac * comprlog + bc;
-    } else {
-        comprthlog = 1.f;
-    }
-
-    if (settings->verbose) {
-        printf("Base Log encoding std=%5.1f\n", (double) linbase);
-    }
-
-    const auto apply =
-    [ = ](float x, bool scale = true) -> float {
-        if (scale)
-        {
-            x /= 65535.f;
-        }
-
-        x = rtengine::max(x, noise);
-        x = rtengine::max(x / gray, noise);
-        if (comprlog && x >= comprthlog)
-        {
-                x = intp(comprfactorlog, (std::tanh((x - comprthlog) / comprthlog) + 1.f) * comprthlog, x);
-        }
-        
-        
-        x = rtengine::max((xlogf(x) / log2 - shadows_range) / dynamic_range, noise);
-        assert(x == x);
-
-
-        if (linbase > 0.f)
-        {
-            x = xlog2lin(x, linbase);
-        }
-
-        if (scale)
-        {
-            return x * 65535.f;
-        } else
-        {
-            return x;
-        }
-    };
-
-    const auto sf =
-        [=](float s, float c) -> float
-        {
-            if (c > noise) {
-
-                return 1.f - min(std::abs(s) / c, 1.f);
-            } else {
-                return 0.f;
-            }
-        };
-//added 2024 02 
-    const auto apply_sat =
-        [&](float &r, float &g, float &b, float f) -> void
-        {
-            float ll = Color::rgbLuminance(r, g, b, ws);
-            float rl = r - ll;
-            float gl = g - ll;
-            float bl = b - ll;
-            float s = intp(max(sf(rl, r), sf(gl, g), sf(bl, b)), pow_F(f, 0.3f) * 0.6f + 0.4f, 1.f);
-            r = ll + s * rl;
-            g = ll + s * gl;
-            b = ll + s * bl;
-        };
-
-
-    float detail = lp.detail;//Log encoding
-    if(lp.cieena) {//Cam16
-        detail = lp.detailcie;
-    }
-    const int W = rgb->getWidth(), H = rgb->getHeight();
-
-    if (detail == 0.f) {//no local contrast
-#ifdef _OPENMP
-        #pragma omp parallel for if (multiThread)
-#endif
-
-        for (int y = 0; y < H; ++y) {
-            for (int x = 0; x < W; ++x) {
-                float r = rgb->r(y, x);
-                float g = rgb->g(y, x);
-                float b = rgb->b(y, x);
-                float m = norm2(r, g, b, ws);
-
-                if (m > noise) {
-                    float mm = apply(m);
-                    float f = mm / m;
-                    f = min(f, 1000000.f);
-
-                    r *= f;
-                    b *= f;
-                    g *= f;
-                    
-                    if (satcontrol && f < 1.f) {
-                        apply_sat(r, g, b, f);
-                    }                    
-                    
-                    r = CLIP(r);
-                    g = CLIP(g);
-                    b = CLIP(b);
-                }
-
-                assert(r == r);
-                assert(g == g);
-                assert(b == b);
-
-                rgb->r(y, x) = r;
-                rgb->g(y, x) = g;
-                rgb->b(y, x) = b;
-            }
-        }
-    } else  {//local contrast
-
-        array2D<float> Y(W, H);
-        {
-            constexpr float base_posterization = 20.f;
-            array2D<float> Y2(W, H);
-
-#ifdef _OPENMP
-            #pragma omp parallel for if (multiThread)
-#endif
-
-            for (int y = 0; y < H; ++y) {
-                for (int x = 0; x < W; ++x) {
-                    Y2[y][x] = norm2(rgb->r(y, x), rgb->g(y, x), rgb->b(y, x), ws) / 65535.f;
-                    float l = xlogf(rtengine::max(Y2[y][x], 1e-9f));
-                    float ll = round(l * base_posterization) / base_posterization;
-                    Y[y][x] = xexpf(ll);
-                    assert(std::isfinite(Y[y][x]));
-                }
-            }
-
-            const float radius = rtengine::max(rtengine::max(bfw, W), rtengine::max(bfh, H)) / 30.f;
-            const float epsilon = 0.005f;
-            rtengine::guidedFilter(Y2, Y, Y, radius, epsilon, multiThread);
-        }
-        const float blend = detail;
-
-#ifdef _OPENMP
-        #pragma omp parallel for if (multiThread)
-#endif
-
-        for (int y = 0; y < H; ++y) {
-            for (int x = 0; x < W; ++x) {
-                float &r = rgb->r(y, x);
-                float &g = rgb->g(y, x);
-                float &b = rgb->b(y, x);
-                float t = Y[y][x];
-                float t2;
-
-                if (t > noise && (t2 = norm2(r, g, b, ws)) > noise) {
-                    float c = apply(t, false);
-                    float f = c / t;
-                    //   float t2 = norm(r, g, b);
-                    float f2 = apply(t2) / t2;
-                    f = intp(blend, f, f2);
-                    f = min(f, 1000000.f);
-
-                    //     assert(std::isfinite(f));
-                    r *= f;
-                    g *= f;
-                    b *= f;
-                    r = CLIP(r);
-                    g = CLIP(g);
-                    b = CLIP(b);
-                    assert(std::isfinite(r));
-                    assert(std::isfinite(g));
-                    assert(std::isfinite(b));
-                    
-                    if (satcontrol && f < 1.f) {
-                        apply_sat(r, g, b, f);
-                    }
-                    
-                    
-                }
-            }
-        }
-
-    }
-}
- 
-// Copyright 2018 Alberto Griggio <alberto.griggio@gmail.com>
-void ImProcFunctions::getAutoLogloc(int sp, ImageSource *imgsrc, float *sourceg, float *blackev, float *whiteev, bool *blackredu,  bool *Autogr, float *sourceab,  int *whits,  int *blacks, int *whitslog,  int *blackslog, int fw, int fh, float xsta, float xend, float ysta, float yend, int SCALE)
-{
-    //BENCHFUN
-//adpatation to local adjustments Jacques Desmis 12 2019 and 11 2021 (from ART)
-// improvment white aand black toen_eqcam 9 2023 
-    const PreviewProps pp(0, 0, fw, fh, SCALE);
-
-    Imagefloat img(int(fw / SCALE + 0.5), int(fh / SCALE + 0.5));
-    const ProcParams neutral;
-
-    imgsrc->getImage(imgsrc->getWB(), TR_NONE, &img, pp, params->toneCurve, neutral.raw);
-    imgsrc->convertColorSpace(&img, params->icm, imgsrc->getWB());
-    float minVal = RT_INFINITY;
-    float maxVal = -RT_INFINITY;
-    TMatrix ws = ICCStore::getInstance()->workingSpaceMatrix(params->icm.workingProfile);
-
-    constexpr float noise = 1e-5;
-    const int h = fh / SCALE;
-    const int w = fw / SCALE;
-
-    const int hsta = ysta * h;
-    const int hend = yend * h;
-
-    const int wsta = xsta * w;
-    const int wend = xend * w;
-    int www = int(fw / SCALE + 0.5);
-    int hhh = int(fh / SCALE + 0.5);
-    array2D<float> YY(www, hhh);
-    double mean = 0.0;
-    int nc = 0;     
-
-    int whit = -whits[sp];
-    int blac = -blacks[sp];
-
-    if(params->locallab.spots.at(sp).expcie && params->locallab.spots.at(sp).Autograycie) {
-        ImProcFunctions::tone_eqcam2(this, &img, whit, blac, params->icm.workingProfile, SCALE, multiThread);
-    }
-
-    int whitlog = -whitslog[sp];
-    int blaclog = -blackslog[sp];
-    
-    if(params->locallab.spots.at(sp).explog && params->locallab.spots.at(sp).autocompute) {
-        ImProcFunctions::tone_eqcam2(this, &img, whitlog, blaclog, params->icm.workingProfile, SCALE, multiThread);
-    }
-
-    for (int y = hsta; y < hend; ++y) {
-        for (int x = wsta; x < wend; ++x) {
-            const float r = img.r(y, x), g = img.g(y, x), b = img.b(y, x);
-            YY[y][x] = norm2(r, g, b, ws) / 65535.f;//norm2 to find a best color luminance response in RGB
-            mean += static_cast<double>((float) ws[1][0] * Color::gamma_srgb(r) + (float) ws[1][1] * Color::gamma_srgb(g) + (float) ws[1][2] * Color::gamma_srgb(b));
-            //alternative to fing gray in case of above process does not works
-            nc++;
-        }
-    }
-
-    for (int y = hsta; y < hend; ++y) {
-        for (int x = wsta; x < wend; ++x) {
-            float l = YY[y][x];
-
-            if (l > noise) {
-                minVal = min(minVal, l);
-                maxVal = max(maxVal, l);
-            }
-        }
-    }
-
-
-    if (!blackredu[sp]){//reduces white point when Freeman algo  or Sigmoid
-        maxVal *= 1.5f;  
-    } 
-    if (!blackredu[sp]){//reduces blackpoint when Freeman algo  or Sigmoid
-        minVal *= 0.5f;
-    }
-
-    //E = 2.5*2^EV => e=2.5 depends on the sensor type C=250 e=2.5 to C=330 e=3.3
-    //repartition with 2.5 between 1.45 Light and shadows 0.58 => a little more 0.55...
-    // https://www.pixelsham.com/2020/12/26/exposure-value-measurements/
-    // https://en.wikipedia.org/wiki/Light_meter
-    if (maxVal > minVal) {
-        const float log2 = std::log(2.f);
-        const float dynamic_range = -xlogf(minVal / maxVal) / log2;
-
-        if (settings->verbose) {
-            std::cout << "AutoLog: min = " << minVal << ", max = " << maxVal
-                      << ", Dynamic Range = " << dynamic_range << std::endl;
-        }
-
-        if (Autogr[sp]) {
-            double tot = 0.0;
-            int n = 0;
-            //0.05 0.25 arbitrary values around gray point 0.18 to find a good value as "gray" for "gain"
-            const float gmax = rtengine::min(maxVal / 2.f, 0.25f);
-            const float gmin = rtengine::max(minVal * std::pow(2.f, rtengine::max((dynamic_range - 1.f) / 2.f, 1.f)), 0.05f);
-
-            if (settings->verbose) {
-                std::cout << "         gray boundaries: " << gmin << ", " << gmax << std::endl;
-            }
-
-            for (int y = hsta; y < hend; ++y) {
-                for (int x = wsta; x < wend; ++x) {
-                    const float l = YY[y][x];
-
-                    if (l >= gmin && l <= gmax) {
-                        tot += static_cast<double>(l);
-                        ++n;
-                    }
-                }
-            }
-
-            if (n > 0) {
-                sourceg[sp] = tot / n * 100.0;
-
-                if (settings->verbose) {
-                    std::cout << "         computed gray point from " << n << " samples: " << sourceg[sp] << std::endl;
-                }
-            } else {//I change slightly this part of algo - more progressivity...best response in very low exposure images
-                mean /= (nc * 65535.0);
-                float yb;
-                yb = 1.5f + 100.f * pow_F(mean, 1.8f);//empirical formula for Jz and log encode for low exposure images
-
-                sourceg[sp] = yb;
-
-                if (settings->verbose) {
-                    std::cout << "         no samples found in range, resorting to Yb gray point value " << sourceg[sp]  << std::endl;
-                }
-            }
-        }
-
-        constexpr float MIN_WHITE = 2.f;
-        constexpr float MAX_BLACK = -3.5f;
-
-        const float gray = sourceg[sp] / 100.f;
-        whiteev[sp] = rtengine::max(xlogf(maxVal / gray) / log2, MIN_WHITE);
-        blackev[sp] = rtengine::min(whiteev[sp] - dynamic_range, MAX_BLACK);
-
-
-        //calculate La - Absolute luminance shooting
-
-        const FramesMetaData* metaData = imgsrc->getMetaData();
-        float fnum = metaData->getFNumber();          // F number
-        float fiso = metaData->getISOSpeed() ;        // ISO
-        float fspeed = metaData->getShutterSpeed() ;  // Speed
-        double fcomp = metaData->getExpComp();        // Compensation +/-
-        double adap;
-
-        if (fnum < 0.3f || fiso < 5.f || fspeed < 0.00001f) { //if no exif data or wrong
-            adap = 2000.;
-        } else {
-            double E_V = fcomp + std::log2(double ((fnum * fnum) / fspeed / (fiso / 100.f)));
-            double kexp = 0.;
-            E_V += kexp * params->toneCurve.expcomp;// exposure compensation in tonecurve ==> direct EV
-            E_V += 0.5 * std::log2(params->raw.expos);  // exposure raw white point ; log2 ==> linear to EV
-            adap = pow(2.0, E_V - 3.0);  // cd / m2  ==> 3.0 = log2(8) =>fnum*fnum/speed = Luminance (average scene) * fiso / K (K is the reflected-light meter calibration constant according to the sensors about 12.5 or 14
-            // end calculation adaptation scene luminosity
-        }
-
-        sourceab[sp] = adap;
-
-    }
 }
 
 
@@ -2799,25 +2332,6 @@ void tone_eq(ImProcFunctions *ipf, Imagefloat *rgb, const struct local_params &l
     params.regularization = lp.detailsh;
     params.pivot = lp.tePivot;
     std::copy(lp.mullocsh, lp.mullocsh + params.bands.size(), params.bands.begin());
-    ipf->toneEqualizer(rgb, params, workingProfile, scale, multithread);
-}
-
-void ImProcFunctions::tone_eqcam(ImProcFunctions *ipf, Imagefloat *rgb, int midtone, const Glib::ustring &workingProfile, double scale, bool multithread)
-{
-    ToneEqualizerParams params;
-    params.enabled = true;
-    params.regularization = 0.f;
-    params.pivot = 0.f;
-    params.bands[0] = 0;
-    params.bands[2] = midtone;
-    params.bands[4] = 0;
-    params.bands[5] = 0;
-    int mid = abs(midtone);
-    int threshmid = 50;
-    if(mid > threshmid) {
-        params.bands[1] = sign(midtone) * (mid - threshmid);
-        params.bands[3] = sign(midtone) * (mid - threshmid);     
-    }
     ipf->toneEqualizer(rgb, params, workingProfile, scale, multithread);
 }
 
@@ -2867,35 +2381,6 @@ void tone_eqsmooth(ImProcFunctions *ipf, Imagefloat *rgb, const struct local_par
     ipf->toneEqualizer(rgb, params, workingProfile, scale, multithread);
 }
 
-void ImProcFunctions::tone_eqcam2(ImProcFunctions *ipf, Imagefloat *rgb, int whits, int blacks, const Glib::ustring &workingProfile, double scale, bool multithread)
-{
-    ToneEqualizerParams params;
-    params.enabled = true;
-    params.regularization = 0.f;
-    params.pivot = 0.f;
-    params.bands[0] = blacks;
-    int bla = abs(blacks);
-    int threshblawhi = 50;
-    int threshblawhi2 = 70;
-    int threshblawhi3 = 40;
-    if(bla > threshblawhi) {
-        params.bands[1] = sign(blacks) * (bla - threshblawhi);
-    }
-    if(bla > threshblawhi2) {
-        params.bands[2] = sign(blacks) * (bla - threshblawhi2);
-    }
-
-    params.bands[4] = whits;
-    int whi = abs(whits);
-    if(whi > threshblawhi) {
-        params.bands[3] = sign(whits) * (whi - threshblawhi);
-    }
-    if(whi > threshblawhi3) {
-        params.bands[5] = sign(whits) * (whi - threshblawhi3);
-    }
-    ipf->toneEqualizer(rgb, params, workingProfile, scale, multithread);
-}
-
 
 void tone_eqblack(ImProcFunctions *ipf, Imagefloat *rgb, int blacks, const Glib::ustring &workingProfile, double scale, bool multithread)
 {
@@ -2907,2501 +2392,20 @@ void tone_eqblack(ImProcFunctions *ipf, Imagefloat *rgb, int blacks, const Glib:
     ipf->toneEqualizer(rgb, params, workingProfile, scale, multithread);
 }
 
+// sigmoidla, gamutjz moved to locallab/iplocallab_ciecam.cc
 
-// tone mapping from
-//  https://github.com/thatcherfreeman/utility-dctls/
-// Copyright of the original code
-/*
-MIT License
+// tone_eqcam, tone_eqcam2, rolloff_freeman_function, scene_referred_contrast,
+// get_freeman_parameters, tonemapFreemanQ, tonemapFreeman, loccont, tone_eqdehaz
+// moved to locallab/iplocallab_tonemap.cc
 
-Copyright (c) 2023 Thatcher Freeman
+// ciecamloc_02float moved to locallab/iplocallab_ciecam.cc
+// (includes helper functions: find_gray, sigmoidla, gamutjz)
 
-Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
+// softproc, softprocess moved to locallab/iplocallab_softlight.cc
 
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
+// exlabLocal moved to locallab/iplocallab_exposure.cc
 
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-*/
-
-/* 
-// I also took some code from Alberto Grigio
-*/
-//Copyright (c) 2023 Thatcher Freeman
-// Adapted to Rawtherapee Jacques Desmis mars / june 2024  jdesmis@gmail.com
-
-float ImProcFunctions::rolloff_freeman_function(float x, float dr, float b, float c, float kmid)
-{
-    return (dr * (x / (x + b)) + c) * kmid;//Simple sigmoid (rather a polynomial asymptotic power function) ponderate with kmid - take into account if need Mean Yb scene and Mean Yb viewing and slope value
-}
-//Copyright (c) 2023 Thatcher Freeman
-// Adapted to Rawtherapee Jacques Desmis mars 2024  jdesmis@gmail.com
-float ImProcFunctions::scene_referred_contrast(float x, float mid_gray_scene, float gamma)
-{
-    return mid_gray_scene * std::pow(x / mid_gray_scene, gamma);//apply gamma
-}
-//Copyright (c) 2023 Thatcher Freeman
-// Adapted to Rawtherapee Jacques Desmis mars 2024  jdesmis@gmail.com
-float ImProcFunctions::get_freeman_parameters(float x, bool rolloff_, float mid_gray_scene, float gamma, float slopelim, float dr, float b, float c, float kmid)
-{
-    if (rolloff_ && x <= mid_gray_scene / slopelim) {//general smooth - till Yb scene
-        return x;
-    } else {
-        return rolloff_freeman_function(scene_referred_contrast(x, mid_gray_scene / slopelim, gamma), dr, b, c, kmid);//simulate polynomial power function with a slope to begin
-    }
-}
-
-//Copyright (c) 2023 Thatcher Freeman
-// Adapted to Rawtherapee Jacques Desmis - August 2024
-
-void ImProcFunctions::tonemapFreemanQ(float Q, float &Qout, float target_slope, float white_point, float black_point, float mid_gray_scene, float mid_gray_view, bool rolloff, bool takeyb)
-{
-    float dr;//Dynamic Range
-    float b;
-    float c;//black point
-    float gamma;
-    float mid_gray_scene_;//Mean luminance - Scene conditions
-    c = black_point;
-    dr = white_point - c;
-    float kmid = 1.f;
-    
-    if(takeyb){
-        kmid = mid_gray_scene / mid_gray_view;
-        kmid = cbrt(kmid);
-    }
-    mid_gray_scene_ = mid_gray_scene;
-    
-    b = (dr / (mid_gray_scene_ - c)) * (1.f - ((mid_gray_scene_ - c) / dr)) * mid_gray_scene_;//b - ponderate mid_gray_scene taking into account the total DR, and the dark part below the mid_gray_scene
-    gamma = target_slope * (float) std::pow((mid_gray_scene_ + b), 2.0) / (dr * b);//Calculate gamma with slope and mid_gray_scene 
-    Qout = get_freeman_parameters(Q, rolloff, mid_gray_scene_, gamma, 1.f, dr, b, c, kmid);//call main function
-}
-
-
-
-// Adapted to Rawtherapee Jacques Desmis 25 mars  - 5 june 2024
-void ImProcFunctions::tonemapFreeman(float target_slope, float target_sloper, float target_slopeg , float target_slopeb, float white_point, float black_point, float mid_gray_scene, float mid_gray_view, bool rolloff, float smooththreshold, bool limslope, LUTf& lut, LUTf& lutr, LUTf& lutg, LUTf& lutb, int mode, bool scale, bool takeyb)
-{
-    float dr;//Dynamic Range
-    float b;
-    float c;//black point
-    float gamma;
-    float gammar;
-    float gammag;
-    float gammab;
-    float mid_gray_scene_;//Mean luminance - Scene conditions // mid_gray_view //Mean luminance - Viewing conditions
-    
-    c = black_point;
-    dr = white_point - c;
-
-    if(scale) {//scale Yb mean luminance scene with white : dr and black
-        mid_gray_scene_ = mid_gray_scene * dr + c;
-    } else {
-        mid_gray_scene_ = mid_gray_scene;
-    }
-
-    b = (dr / (mid_gray_scene_ - c)) * (1.f - ((mid_gray_scene_ - c) / dr)) * mid_gray_scene_;//b - ponderate mid_gray_scene taking into account the total DR, and the dark part below the mid_gray_scene
-    gamma = target_slope * (float) std::pow((mid_gray_scene_ + b), 2.0) / (dr * b);//Caculate gamma with slope and mid_gray_scene 
-    gammar = target_sloper * (float) std::pow((mid_gray_scene_ + b), 2.0) / (dr * b);//Caculate gamma with slope and mid_gray_scene 
-    gammag = target_slopeg * (float) std::pow((mid_gray_scene_ + b), 2.0) / (dr * b);//Caculate gamma with slope and mid_gray_scene 
-    gammab = target_slopeb * (float) std::pow((mid_gray_scene_ + b), 2.0) / (dr * b);//Caculate gamma with slope and mid_gray_scene 
-    float kmid = 1.f;//general case
-    //float kyb = 1.f;
-    if(takeyb){
-        kmid = mid_gray_scene / mid_gray_view;
-        kmid = cbrt(kmid);
-    }
-   // if(mode == 3 && target_slope != 1.f ) {//case tone-mapping
-/*
-         
-        float midutil = mid_gray_view / mid_gray_scene;//take into account ratio between Yb source and Yb viewing
-        float midk = 1.f;
-        float k_slope = 2.2f;
-        if(target_slope >= 1.f) {
-            midk = pow_F(midutil, k_slope * (target_slope - 1.f));//ponderation in function target_slope when "slope user" < 1.f
-        }
-        kmid = midk;
-        
-    }
-*/    
-    if (mode == 3 && settings->verbose) {
-        printf("b=%f gamma=%f slope=%f DynRange=%f kmid=%f black=%f Yb-scale=%f\n", (double) b, (double) gamma, (double) target_slope, (double) dr, (double) kmid, (double) c, (double) mid_gray_scene_);
-    }
-    //lut - take from Alberto Griggio
-    if(mode == 4) {
-        float sloplimr = 1.f;
-        float sloplimg = 1.f;
-        float sloplimb = 1.f;
-        if(limslope) {
-            rolloff = true;
-        }  
-        //always apply threshold
-        sloplimr *= smooththreshold;
-        sloplimg *= smooththreshold;
-        sloplimb *= smooththreshold;
-        
-        for (int i = 0; i < 65536; ++i) {// i - value image RGB
-            lutr[i] = get_freeman_parameters(float(i) / 65535.f, rolloff, mid_gray_scene_, gammar, sloplimr, dr, b, c, kmid);//call main function
-            lutg[i] = get_freeman_parameters(float(i) / 65535.f, rolloff, mid_gray_scene_, gammag, sloplimg, dr, b, c, kmid);//call main function
-            lutb[i] = get_freeman_parameters(float(i) / 65535.f, rolloff, mid_gray_scene_, gammab, sloplimb, dr, b, c, kmid);//call main function
-        }
-    } else {
-        kmid = 1.f;
-        for (int i = 0; i < 65536; ++i) {// i - value image RGB
-            lut[i] = get_freeman_parameters(float(i) / 65535.f, rolloff, mid_gray_scene_, gamma, 1.f, dr, b, c, kmid);//call main function
-        }
-    }
-}
-
-
-void ImProcFunctions::loccont(int bfw, int bfh, LabImage* tmp1, float rad, float stren, int sk)
-{
-    if (rad > 0.f) {
-        array2D<float> guide(bfw, bfh);
-        array2D<float> LL(bfw, bfh);
-#ifdef _OPENMP
-        #pragma omp parallel for schedule(dynamic,16)
-#endif
-
-        for (int y = 0; y < bfh ; y++) {
-            for (int x = 0; x < bfw; x++) {
-                LL[y][x] = tmp1->L[y][x];
-                float ll = LL[y][x] / 32768.f;
-                guide[y][x] = xlin2log(rtengine::max(ll, 0.f), 10.f);
-            }
-        }
-
-        array2D<float> iL(bfw, bfh, LL, 0);
-        float gu = stren * rad;
-        int r = rtengine::max(int(gu / sk), 1);
-        const double epsil = 0.001 * std::pow(2.f, -10);
-        float st = 0.01f * rad;
-        rtengine::guidedFilterLog(guide, 10.f, LL, r, epsil, false);
-
-#ifdef _OPENMP
-        #pragma omp parallel for schedule(dynamic,16)
-#endif
-
-        for (int y = 0; y < bfh ; y++) {
-            for (int x = 0; x < bfw; x++) {
-                LL[y][x] = intp(st, LL[y][x], iL[y][x]);
-                tmp1->L[y][x] = LL[y][x];
-            }
-        }
-    }
-}
-
-void ImProcFunctions::tone_eqdehaz(ImProcFunctions *ipf, Imagefloat *rgb, int whits, int blacks, const Glib::ustring &workingProfile, double scale, bool multithread)
-{
-    ToneEqualizerParams params;
-    params.enabled = true;
-    params.regularization = 0.f;
-    params.pivot = 0.f;
-    double blred = 0.4;
-    params.bands[0] = blred * blacks;
-    int bla = abs(blacks);
-    int threshblawhi = 50;
-    int threshblawhi2 = 85;
-    if(bla > threshblawhi) {
-        params.bands[1] = blred * sign(blacks) * (bla - threshblawhi);
-    }
-    if(bla > threshblawhi2) {
-        params.bands[2] = blred * sign(blacks) * (bla - threshblawhi2);
-    }
-
-    params.bands[4] = whits;
-    int whi = abs(whits);
-    if(whi > threshblawhi) {
-        params.bands[3] = sign(whits) * (whi - threshblawhi);
-    }
-    ipf->toneEqualizer(rgb, params, workingProfile, scale, multithread);
-}
-
-
-void sigmoidla(float &valj, float thresj, float lambda)
-{
-    //thres : shifts the action of sigmoid to darker tones or lights
-    //lambda : changes the "slope" of the sigmoid. Low values give a flat curve, high values a "rectangular / orthogonal" curve
-    valj =  1.f / (1.f + xexpf(lambda - (lambda / thresj) * valj));
-}
-
-
-void gamutjz(double &Jz, double &az, double &bz, double pl, const double wip[3][3], const float higherCoef, const float lowerCoef)
-{
-    //Not used...bad results
-    constexpr float ClipLevel = 65535.0f;
-    bool inGamut;
-
-    //  int nb = 0;
-    do {
-        inGamut = true;
-        double L_, M_, S_;
-        double xx, yy, zz;
-        bool zcam = false;
-        Ciecam02::jzczhzxyz(xx, yy, zz, Jz, az, bz, pl, L_, M_, S_, zcam);
-        double x, y, z;
-        x = 65535. * (d65_d50[0][0] * xx + d65_d50[0][1] * yy + d65_d50[0][2] * zz);
-        y = 65535. * (d65_d50[1][0] * xx + d65_d50[1][1] * yy + d65_d50[1][2] * zz);
-        z = 65535. * (d65_d50[2][0] * xx + d65_d50[2][1] * yy + d65_d50[2][2] * zz);
-        float R, G, B;
-        Color:: xyz2rgb(x, y, z, R, G, B, wip);
-
-        if (rtengine::min(R, G, B) < 0.f  || rtengine::max(R, G, B) > ClipLevel) {
-            //    nb++;
-            double hz = xatan2f(bz, az);
-            float2 sincosval = xsincosf(hz);
-            double Cz = sqrt(az * az + bz * bz);
-            Cz *= (double) higherCoef;
-
-            if (Cz < 0.01 && Jz > 0.05) { //empirical values
-                Jz -= (double) lowerCoef;
-            }
-
-            az = clipazbz(Cz * (double) sincosval.y);
-            bz = clipazbz(Cz * (double) sincosval.x);
-
-            inGamut = false;
-        }
-    } while (!inGamut);
-}
-
-void ImProcFunctions::ciecamloc_02float(struct local_params& lp, int sp, LabImage* lab, int bfw, int bfh, int call, int sk, const LUTf& cielocalcurve, bool localcieutili, const LUTf& cielocalcurve2, bool localcieutili2,
-                                        const LUTf& jzlocalcurve, bool localjzutili, const LUTf& czlocalcurve, bool localczutili, const LUTf& czjzlocalcurve, bool localczjzutili, const LocCHCurve& locchCurvejz, const LocHHCurve& lochhCurvejz, const LocLHCurve& loclhCurvejz, bool HHcurvejz, bool CHcurvejz, bool LHcurvejz,
-                                        const LocwavCurve& locwavCurvejz, bool locwavutilijz, float &maxicam, float &contsig, float &lightsig
-                                       )
-{
-//    BENCHFUN
-    if (!params->locallab.spots.at(sp).activ) { //disable all ciecam functions
-        return;
-    }
-
-    bool ciec = false;
-    bool iscie = false;
-
-
-    int modeqj = 1;
-    if (params->locallab.spots.at(sp).modeQJ == "511") {
-        modeqj = 0;
-    } else if (params->locallab.spots.at(sp).modeQJ == "512") {
-        modeqj = 1;
-    }
-
-    if (params->locallab.spots.at(sp).ciecam && params->locallab.spots.at(sp).explog && call == 1) {
-        ciec = true;
-        iscie = false;
-    } else if (params->locallab.spots.at(sp).expcie && call == 0) {
-        ciec = true;
-        iscie = true;
-    }
-
-    bool z_cam = false; //params->locallab.spots.at(sp).jabcie; //alaways use normal algorithm, Zcam giev often bad results
-    bool jabcie = false;//always disabled
-    bool issigjz12 = params->locallab.spots.at(sp).sigjz12;
-    bool issigq12 = params->locallab.spots.at(sp).sigq12;
-    
-    bool islogjz = params->locallab.spots.at(sp).forcebw;
-    bool issigjz = params->locallab.spots.at(sp).sigjz;
-    bool issigq = params->locallab.spots.at(sp).sigq;
-   
-    bool issig = true; //params->locallab.spots.at(sp).sigcie;
-
-    //sigmoid J Q variables
-   // const float sigmoidlambda = params->locallab.spots.at(sp).sigmoidldacie12;
-   // const float sigmoidth = params->locallab.spots.at(sp).sigmoidthcie;
-   // const float sigmoidbl = params->locallab.spots.at(sp).sigmoidblcie12;
-    const bool sigmoidnorm = params->locallab.spots.at(sp).normcie;
-
-    const float sigmoidlambda = params->locallab.spots.at(sp).sigmoidldacie;
-    const float sigmoidth = params->locallab.spots.at(sp).sigmoidthcie;
-    const float sigmoidbl = params->locallab.spots.at(sp).sigmoidblcie;
-
-
-    int mobwev12 = 0;
-    if (params->locallab.spots.at(sp).bwevMethod12 == "sigQ") {
-        mobwev12 = 0;
-    } else if (params->locallab.spots.at(sp).bwevMethod12 == "slop") {
-        mobwev12 = 1;
-    }
-
-    int mobwev = 0;
-    float sumcamq01 = 0.5f;
-
-    if (params->locallab.spots.at(sp).bwevMethod == "none") {
-        mobwev = 0;
-    } else if (params->locallab.spots.at(sp).bwevMethod == "sig") {
-        mobwev = 1;
-    } else if (params->locallab.spots.at(sp).bwevMethod == "logsig") {
-        mobwev = 2;
-    }
-
-    float senssig =(float) params->locallab.spots.at(sp).sigmoidsenscie;
-
-    float middle_grey_contrast = params->locallab.spots.at(sp).sigmoidldacie12;
-    float contrast_skewness = params->locallab.spots.at(sp).sigmoidthcie12;
-    float white_point_disp = params->locallab.spots.at(sp).sigmoidblcie12;
-    float middle_grey = 0.01 * params->locallab.spots.at(sp).sourceGraycie;
-    middle_grey *= 2.f;//take into account Ciecam
-    middle_grey = std::min(middle_grey, 0.6f);
-
-    float black_point =  xexpf(lp.blackevjz * std::log(2.f) + xlogf(middle_grey));
-    float white_pointsig = xexpf(lp.whiteevjz * std::log(2.f) + xlogf(middle_grey));//to adapt if need and remove slider whitsig
-
-    float slopsmootq =(float) params->locallab.spots.at(sp).slopesmoq;
-    float mid_gray_view = 0.01f * lp.targetgraycie;
-    TMatrix wiprof = ICCStore::getInstance()->workingSpaceInverseMatrix(params->icm.workingProfile);
-    const double wip[3][3] = {//improve precision with double
-        {wiprof[0][0], wiprof[0][1], wiprof[0][2]},
-        {wiprof[1][0], wiprof[1][1], wiprof[1][2]},
-        {wiprof[2][0], wiprof[2][1], wiprof[2][2]}
-    };
-    float plum = (float) params->locallab.spots.at(sp).pqremapcam16;
-
-    int mocam = 1;
- 
-    if(lp.moka == 1) {
-        mocam = 1;
-    } else if (lp.moka == 2) {
-        mocam = 2;
-    }
-
-    int mecamcurve = 0;
-
-    if (params->locallab.spots.at(sp).toneMethodcie == "one") {
-        mecamcurve = 0;
-    } else if (params->locallab.spots.at(sp).toneMethodcie == "two") {
-        mecamcurve = 1;
-    }
-
-    int mecamcurve2 = 0;
-
-    if (params->locallab.spots.at(sp).toneMethodcie2 == "onec") {
-        mecamcurve2 = 0;
-    } else if (params->locallab.spots.at(sp).toneMethodcie2 == "twoc") {
-        mecamcurve2 = 1;
-    } else if (params->locallab.spots.at(sp).toneMethodcie2 == "thrc") {
-        mecamcurve2 = 2;
-    }
-
-    float th = 1.f;
-//    const float at = 1.f - sigmoidth;
-//    const float bt = sigmoidth;
-
-   // const float ath = sigmoidth - 1.f;
-   // const float bth = 1;
-    float sila = pow_F(sigmoidlambda, senssig);
-    sila = LIM01(sila);
-    const float sigm = 3.3f + 7.1f * (1.f - sila); //e^10.4 = 32860 => sigm vary from 3.3 to 10.4
-    float bl = std::min(sigmoidbl, 1.f);//reused old slider
-    if(params->locallab.spots.at(sp).logcieq) {
-        bl = 0.01f * (float) params->locallab.spots.at(sp).strcielog;
-        bl = std::min(bl, 1.f);
-    }
-
-    //end sigmoid
-
-    int width = lab->W, height = lab->H;
-    float Yw;
-    Yw = 1.0f;
-    double Xw, Zw;
-    float f = 0.f, nc = 0.f, la, c = 0.f, xw, yw, zw, f2 = 1.f, c2 = 1.f, nc2 = 1.f, yb2;
-    float fl, n, nbb, ncb, aw; //d
-    float xwd, ywd, zwd, xws, yws, zws;
-    //  int alg = 0;
-    double Xwout, Zwout;
-    double Xwsc, Zwsc;
-
-    LUTu hist16J(32768, LUT_CLIP_BELOW | LUT_CLIP_ABOVE, true);
-    LUTu hist16Q(32768, LUT_CLIP_BELOW | LUT_CLIP_ABOVE, true);
-    //for J light and contrast
-    LUTf CAMBrightCurveJ(32768, LUT_CLIP_BELOW | LUT_CLIP_ABOVE);
-    LUTf CAMBrightCurveQ(32768, LUT_CLIP_BELOW | LUT_CLIP_ABOVE);
-    LUTf CAMBrightCurveQsig(32768, LUT_CLIP_BELOW | LUT_CLIP_ABOVE);
-
-#ifdef _OPENMP
-    const int numThreads = min(max(width * height / 65536, 1), omp_get_max_threads());
-    #pragma omp parallel num_threads(numThreads) if(numThreads>1)
-#endif
-    {
-        LUTu hist16Jthr(hist16J.getSize(), LUT_CLIP_BELOW | LUT_CLIP_ABOVE, true);
-        LUTu hist16Qthr(hist16Q.getSize(), LUT_CLIP_BELOW | LUT_CLIP_ABOVE, true);
-
-#ifdef _OPENMP
-        #pragma omp for
-#endif
-
-        for (int i = 0; i < height; i++) {
-            for (int j = 0; j < width; j++) { //rough correspondence between L and J
-                float currL = lab->L[i][j] / 327.68f;
-                float koef; //rough correspondence between L and J
-
-                if (currL > 50.f) {
-                    if (currL > 70.f) {
-                        if (currL > 80.f) {
-                            if (currL > 85.f) {
-                                koef = 0.97f;
-                            } else {
-                                koef = 0.93f;
-                            }
-                        } else {
-                            koef = 0.87f;
-                        }
-                    } else {
-                        if (currL > 60.f) {
-                            koef = 0.85f;
-                        } else {
-                            koef = 0.8f;
-                        }
-                    }
-                } else {
-                    if (currL > 10.f) {
-                        if (currL > 20.f) {
-                            if (currL > 40.f) {
-                                koef = 0.75f;
-                            } else {
-                                koef = 0.7f;
-                            }
-                        } else {
-                            koef = 0.9f;
-                        }
-                    } else {
-                        koef = 1.0;
-                    }
-                }
-
-                hist16Jthr[(int)((koef * lab->L[i][j]))]++;    //evaluate histogram luminance L # J
-                hist16Qthr[CLIP((int)(32768.f * sqrt((koef * (lab->L[i][j])) / 32768.f)))]++;     //for brightness Q : approximation for Q=wh*sqrt(J/100)  J not equal L
-            }
-        }
-
-#ifdef _OPENMP
-        #pragma omp critical
-#endif
-        {
-            hist16J += hist16Jthr;
-            hist16Q += hist16Qthr;
-        }
-    }
-#ifdef _OPENMP
-    static_cast<void>(numThreads); // to silence cppcheck warning
-#endif
-
-    //evaluate lightness, contrast
-
-    if (ciec) {
-        float contL = 0.f;
-        float lightL = 0.f;
-        float contQ = 0.f;
-        float lightQ = 0.f;
-
-        if (iscie) {
-            contL = 0.6 * params->locallab.spots.at(sp).contlcie; //0.6 less effect, no need 1.
-            lightL = 0.4 * params->locallab.spots.at(sp).lightlcie; //0.4 less effect, no need 1.
-            contQ = 0.5 * params->locallab.spots.at(sp).contqcie; //0.5 less effect, no need 1.
-            lightQ = 0.4 * params->locallab.spots.at(sp).lightqcie; //0.4 less effect, no need 1.
-        } else {
-            contL = 0.6 * params->locallab.spots.at(sp).contl; //0.6 less effect, no need 1.
-            lightL = 0.4 * params->locallab.spots.at(sp).lightl; //0.4 less effect, no need 1.
-            contQ = 0.5 * params->locallab.spots.at(sp).contq; //0.5 less effect, no need 1.
-            lightQ = 0.4 * params->locallab.spots.at(sp).lightq; //0.4 less effect, no need 1.
-
-        }
-
-        float contthresL = 0.f;
-
-        if (iscie) {
-            contthresL = params->locallab.spots.at(sp).contthrescie;
-        } else {
-            contthresL = params->locallab.spots.at(sp).contthres;
-        }
-
-        float contthresQ = contthresL;
-
-        if (contL < 0.f) {
-            contthresL *= -1;
-        }
-
-        float thL = 0.6f;
-        thL = 0.3f * contthresL + 0.6f;
-
-        if (contQ < 0.f) {
-            contthresQ *= -1;
-        }
-
-        float thQ = 0.6f;
-        thQ = 0.3f * contthresQ + 0.6f;
-        float thQsig = 0.6f;
-        Ciecam02::curveJfloat(lightL, contL, thL, hist16J, CAMBrightCurveJ); //lightness J and contrast J
-        CAMBrightCurveJ /= 327.68f;
-        double podcont = 40.;//50
-        double podcont0 = 40.;//50
-        double podcont1 = 30.;//35.
-        double ka = -(podcont0 - podcont1) / 0.5;
-        double kb = podcont1 - ka;
-
-        double podlight = 35.;
-        double podlight0 = 35.;
-        double podlight1 = 40.;//45
-        double kal = -(podlight0 - podlight1) / 0.5;
-        double kbl = podlight1 - kal;
-        
-        double contbase = params->locallab.spots.at(sp).sigmoidldacie;
-        
-        if(contbase <= 0.5)  {
-            podcont = podcont0;
-            podlight = podlight0;
-            
-        } else {
-            podcont = ka * contbase + kb;
-            podlight = kal * contbase + kbl;           
-        }
-        Ciecam02::curveJfloat(lightQ, contQ, thQ, hist16Q, CAMBrightCurveQ); //brightness Q and contrast Q
-        lightsig = -podlight * contbase;
-        contsig = podcont * contbase;
-        Ciecam02::curveJfloat(lightsig, contsig, thQsig, hist16Q, CAMBrightCurveQsig); //brightness Q and contrast Q bypass.
-    }
-
-
-    int tempo = 5000;
-
-    if (params->locallab.spots.at(sp).expvibrance && call == 2) {
-        if (params->locallab.spots.at(sp).warm > 0) {
-            tempo = 5000 - 30 * params->locallab.spots.at(sp).warm;
-        } else if (params->locallab.spots.at(sp).warm < 0) {
-            tempo = 5000 - 70 * params->locallab.spots.at(sp).warm;
-        }
-    }
-
-
-    if (ciec) {
-        if (iscie) {
-            if (params->locallab.spots.at(sp).catadcie > 0) {
-                tempo = 5000 - 30 * params->locallab.spots.at(sp).catadcie;
-            } else if (params->locallab.spots.at(sp).catadcie < 0) {
-                tempo = 5000 - 70 * params->locallab.spots.at(sp).catadcie;
-            }
-        } else {
-            if (params->locallab.spots.at(sp).catad > 0) {
-                tempo = 5000 - 30 * params->locallab.spots.at(sp).catad;
-            } else if (params->locallab.spots.at(sp).catad < 0) {
-                tempo = 5000 - 70 * params->locallab.spots.at(sp).catad;
-            }
-        }
-    }
-
-    ColorTemp::temp2mulxyz(params->wb.temperature, params->wb.method, params->wb.observer, Xw, Zw);  //compute white Xw Yw Zw  : white current WB
-    ColorTemp::temp2mulxyz(tempo, "Custom", params->wb.observer, Xwout, Zwout);
-    ColorTemp::temp2mulxyz(5000, "Custom", params->wb.observer, Xwsc, Zwsc);
-
-    //viewing condition for surrsrc
-    f  = 1.00f;
-    c  = 0.69f;
-    nc = 1.00f;
-    //viewing condition for surround
-    f2 = 1.0f, c2 = 0.69f, nc2 = 1.0f;
-
-    if (ciec) {
-        if (iscie) {
-            //surround source with only 2 choices (because Log encoding before)
-            if(lp.sursouci == 0) {
-                f = 1.0f, c = 0.69f, nc = 1.0f;
-            } else if (lp.sursouci == 1){
-                f  = 0.9f;
-                c  = 0.59f;
-                nc = 0.9f;
-            } else if (lp.sursouci == 2) {
-                f  = 0.8f;
-                c  = 0.525f;
-                nc = 0.8f;
-            } else if (lp.sursouci == 3) {
-                f  = 0.8f;
-                c  = 0.41f;
-                nc = 0.8f;
-            } else if (lp.sursouci == 4) {
-                f = 1.0f, c = 0.702f, nc = 1.0f;//very small surround effect for Jz - Also disable Ciecam further
-            }
-        } else {
-            if (params->locallab.spots.at(sp).sursour == "Average") {
-                f = 1.0f, c = 0.69f, nc = 1.0f;
-            } else if (params->locallab.spots.at(sp).sursour == "Dim") {
-                f  = 0.9f;
-                c  = 0.59f;
-                nc = 0.9f;
-            } else if (params->locallab.spots.at(sp).sursour == "Dark") {
-                f  = 0.8f;
-                c  = 0.525f;
-                nc = 0.8f;
-            } else if (params->locallab.spots.at(sp).sursour == "exDark") {
-                f  = 0.8f;
-                c  = 0.41f;
-                nc = 0.8f;
-            }
-        }
-
-        //viewing condition for surround
-        if (iscie) {
-            if (params->locallab.spots.at(sp).surroundcie == "Average") {
-                f2 = 1.0f, c2 = 0.69f, nc2 = 1.0f;
-            } else if (params->locallab.spots.at(sp).surroundcie == "Dim") {
-                f2  = 0.9f;
-                c2  = 0.59f;
-                nc2 = 0.9f;
-            } else if (params->locallab.spots.at(sp).surroundcie == "Dark") {
-                f2  = 0.8f;
-                c2  = 0.525f;
-                nc2 = 0.8f;
-            } else if (params->locallab.spots.at(sp).surroundcie == "ExtremelyDark") {
-                f2  = 0.8f;
-                c2  = 0.41f;
-                nc2 = 0.8f;
-            }
-        } else {
-            if (params->locallab.spots.at(sp).surround == "Average") {
-                f2 = 1.0f, c2 = 0.69f, nc2 = 1.0f;
-            } else if (params->locallab.spots.at(sp).surround == "Dim") {
-                f2  = 0.9f;
-                c2  = 0.59f;
-                nc2 = 0.9f;
-            } else if (params->locallab.spots.at(sp).surround == "Dark") {
-                f2  = 0.8f;
-                c2  = 0.525f;
-                nc2 = 0.8f;
-            } else if (params->locallab.spots.at(sp).surround == "ExtremelyDark") {
-                f2  = 0.8f;
-                c2  = 0.41f;
-                nc2 = 0.8f;
-            }
-
-        }
-    }
-
-    xwd = 100.0 * Xwout;
-    zwd = 100.0 * Zwout;
-    ywd = 100.f;
-
-    xws = 100.0 * Xwsc;
-    zws = 100.0 * Zwsc;
-    yws = 100.f;
-
-
-    //La and la2 = ambiant luminosity scene and viewing
-    la = 400.f;
-    float la2 = 400.f;
-
-    if (ciec) {
-        if (iscie) {
-            la = params->locallab.spots.at(sp).sourceabscie;
-            la2 = params->locallab.spots.at(sp).targabscie;
-        } else {
-            la = params->locallab.spots.at(sp).sourceabs;
-            la2 = params->locallab.spots.at(sp).targabs;
-        }
-    }
-
-    const float pilot = 2.f;
-    const float pilotout = 2.f;
-    double avgm = 0.;
-    //algoritm's params
-    float yb = 18.f;
-    yb2 = 18;
-
-    if (ciec) {
-        if (iscie) {
-            yb = params->locallab.spots.at(sp).sourceGraycie;//
-            avgm = (double) pow_F(0.01f * (yb - 1.f), 0.45f);;
-            yb2 = params->locallab.spots.at(sp).targetGraycie;
-        } else {
-            yb = params->locallab.spots.at(sp).targetGray;//target because we are after Log encoding
-            yb2 = params->locallab.spots.at(sp).targetGray;
-        }
-    }
-
-    if (params->locallab.spots.at(sp).expcie && call == 10 && params->locallab.spots.at(sp).modecam == "jz") {
-        yb = params->locallab.spots.at(sp).sourceGraycie;//for Jz calculate Yb and surround in Lab and cam16 before process Jz
-        la = params->locallab.spots.at(sp).sourceabscie;
-            if(lp.sursouci == 0) {
-                f = 1.0f, c = 0.69f, nc = 1.0f;
-            } else if (lp.sursouci == 1){
-                f  = 0.9f;
-                c  = 0.59f;
-                nc = 0.9f;
-            } else if (lp.sursouci == 2) {
-                f  = 0.8f;
-                c  = 0.525f;
-                nc = 0.8f;
-            } else if (lp.sursouci == 3) {
-                f  = 0.8f;
-                c  = 0.41f;
-                nc = 0.8f;
-            } else if (lp.sursouci == 4) {
-                f = 1.0f, c = 0.702f, nc = 1.0f;//very small surround effect for Jz
-            }
-       
-    }
-
-    float schr = 0.f;
-    float mchr = 0.f;
-    float cchr = 0.f;
-    float rstprotection = 0.f;
-    float hue = 0.f;
-
-    if (ciec) {
-        if (iscie) {
-            rstprotection =  params->locallab.spots.at(sp).rstprotectcie;
-            hue = params->locallab.spots.at(sp).huecie;
-
-            cchr = params->locallab.spots.at(sp).chromlcie;
-
-            if (cchr == -100.0f) {
-                cchr = -99.8f;
-            }
-
-            schr = params->locallab.spots.at(sp).saturlcie;
-
-            if (schr > 0.f) {
-                schr = schr / 2.f;    //divide sensibility for saturation
-            }
-
-            if (schr == -100.f) {
-                schr = -99.8f;
-            }
-
-            mchr = params->locallab.spots.at(sp).colorflcie;
-
-            if (mchr == -100.0f) {
-                mchr = -99.8f ;
-            }
-
-            if (mchr == 100.0f) {
-                mchr = 99.9f;
-            }
-
-        } else {
-            cchr = params->locallab.spots.at(sp).chroml;
-
-            if (cchr == -100.0f) {
-                cchr = -99.8f;
-            }
-
-            schr = params->locallab.spots.at(sp).saturl;
-
-            if (schr > 0.f) {
-                schr = schr / 2.f;    //divide sensibility for saturation
-            }
-
-            if (schr == -100.f) {
-                schr = -99.8f;
-            }
-
-            mchr = params->locallab.spots.at(sp).colorfl;
-
-            if (mchr == -100.0f) {
-                mchr = -99.8f ;
-            }
-
-            if (mchr == 100.0f) {
-                mchr = 99.9f;
-            }
-        }
-    }
-
-    float d, dj;
-
-    // const int gamu = 0; //(params->colorappearance.gamut) ? 1 : 0;
-    xw = 100.0 * Xw;
-    yw = 100.f * Yw;
-    zw = 100.0 * Zw;
-    float xw1 = xws, yw1 = yws, zw1 = zws, xw2 = xwd, yw2 = ywd, zw2 = zwd;
-    float cz, wh, pfl;
-    int c16 = 16;//always cat16
-    bool c20 = true;
-
-    if (c20  && plum > 100.f) {
-        c16 = 21;//I define 21...for 2021 :)
-    }
-
-    int level_bljz = params->locallab.spots.at(sp).csthresholdjz.getBottomLeft();
-    int level_hljz = params->locallab.spots.at(sp).csthresholdjz.getTopLeft();
-    int level_brjz = params->locallab.spots.at(sp).csthresholdjz.getBottomRight();
-    int level_hrjz = params->locallab.spots.at(sp).csthresholdjz.getTopRight();
-
-    float alowjz = 1.f;
-    float blowjz = 0.f;
-
-    if (level_hljz != level_bljz) {
-        alowjz = 1.f / (level_hljz - level_bljz);
-        blowjz = -alowjz * level_bljz;
-    }
-
-    float ahighjz = 1.f;
-    float bhighjz = 0.f;
-
-    if (level_hrjz != level_brjz) {
-        ahighjz = 1.f / (level_hrjz - level_brjz);
-        bhighjz =  -ahighjz * level_brjz;
-    }
-
-    float sigmalcjz = params->locallab.spots.at(sp).sigmalcjz;
-    float jzamountchr = 0.01 * params->locallab.spots.at(sp).thrhjzcie;
-    bool jzch = params->locallab.spots.at(sp).chjzcie;
-    double jzamountchroma = 0.01 * settings->amchromajz;
-
-    if (jzamountchroma < 0.05) {
-        jzamountchroma = 0.05;
-    }
-
-    if (jzamountchroma > 2.) {
-        jzamountchroma = 2.;
-    }
-
-    Ciecam02::initcam1float(yb, pilot, f, la, xw, yw, zw, n, d, nbb, ncb, cz, aw, wh, pfl, fl, c, c16, plum);
-    const float pow1 = pow_F(1.64f - pow_F(0.29f, n), 0.73f);
-    float nj, nbbj, ncbj, czj, awj, flj;
-    Ciecam02::initcam2float(yb2, pilotout, f2,  la2,  xw2,  yw2,  zw2, nj, dj, nbbj, ncbj, czj, awj, flj, c16, plum);
-#if defined(__SSE2__) || defined(RT_SIMDE)
-    const float reccmcz = 1.f / (c2 * czj);
-#endif
-    const float epsil = 0.0001f;
-    const float coefQ = 32767.f / wh;
-    const float coefq = 1.f / wh;
-
-    const float pow1n = pow_F(1.64f - pow_F(0.29f, nj), 0.73f);
-    const float coe = pow_F(fl, 0.25f);
-    const float QproFactor = (0.4f / c) * (aw + 4.0f) ;
-    const double shadows_range =  params->locallab.spots.at(sp).blackEvjz;
-    const double targetgray = params->locallab.spots.at(sp).targetjz;
-    double targetgraycor = 0.15;
-    double dynamic_range = std::max(params->locallab.spots.at(sp).whiteEvjz - shadows_range, 0.5);
-    const double noise = pow(2., -16.6);//16.6 instead of 16 a little less than others, but we work in double
-    const double log2 = xlog(2.);
-    const float log2f = xlogf(2.f);
-    
-    float middle_grey_contrastjz = params->locallab.spots.at(sp).sigmoidldajzcie12;
-    float contrast_skewnessjz = params->locallab.spots.at(sp).sigmoidthjzcie12;
-    float white_point_dispjz = params->locallab.spots.at(sp).sigmoidbljzcie12;
-    float middle_greyjz = 0.01 * params->locallab.spots.at(sp).sourceGraycie;
-    middle_greyjz *= 2.f;
-    middle_greyjz = std::min(middle_greyjz, 0.6f);
-
-    float black_pointjz =  xexpf(lp.blackevjz * std::log(2.f) + xlogf(middle_greyjz));
-    float white_pointsigjz = xexpf(lp.whiteevjz * std::log(2.f) + xlogf(middle_greyjz));//to adapt if need and remove slider whitsig
-    float drjz = white_pointsigjz - black_pointjz;
-    if(params->locallab.spots.at(sp).sigybjz12) {
-        middle_greyjz = middle_greyjz * drjz + black_pointjz;
-    }
-
-    if ((mocam == 2)  && call == 0) { //Jz az bz ==> Jz Cz Hz before Ciecam16
-        double mini = 1000.;
-        double maxi = -1000.;
-        double sum = 0.;
-        int nc = 0;
-        double epsiljz = 0.0001;
-        //Remapping see https://hal.inria.fr/hal-02131890/document    I took some ideas in this text, and add my personal adaptation
-        // image quality assessment of HDR and WCG images https://tel.archives-ouvertes.fr/tel-02378332/document
-        double adapjz = params->locallab.spots.at(sp).adapjzcie;
-        double jz100 = params->locallab.spots.at(sp).jz100;
-        double pl = params->locallab.spots.at(sp).pqremap;
-        double jzw, azw, bzw;
-        jzw = 0.18;//Jz white
-
-        bool Qtoj = params->locallab.spots.at(sp).qtoj;//betwwen lightness to brightness
-        const bool logjz =  params->locallab.spots.at(sp).logjz;//log encoding
-//calculate min, max, mean for Jz
-#ifdef _OPENMP
-        #pragma omp parallel for reduction(min:mini) reduction(max:maxi) reduction(+:sum) if(multiThread)
-#endif
-
-        for (int i = 0; i < height; i += 1) {
-            for (int k = 0; k < width; k += 1) {
-                float L = lab->L[i][k];
-                float a = lab->a[i][k];
-                float b = lab->b[i][k];
-                float x, y, z;
-                //convert Lab => XYZ
-                Color::Lab2XYZ(L, a, b, x, y, z);
-                x = x / 65535.f;
-                y = y / 65535.f;
-                z = z / 65535.f;
-                double Jz, az, bz;
-                double xx, yy, zz;
-                //D50 ==> D65
-                xx = (d50_d65[0][0] * (double) x + d50_d65[0][1] * (double) y + d50_d65[0][2] * (double) z);
-                yy = (d50_d65[1][0] * (double) x + d50_d65[1][1] * (double) y + d50_d65[1][2] * (double) z);
-                zz = (d50_d65[2][0] * (double) x + d50_d65[2][1] * (double) y + d50_d65[2][2] * (double) z);
-
-                double L_p, M_p, S_p;
-                bool zcam = z_cam;
-
-                Ciecam02::xyz2jzczhz(Jz, az, bz, xx, yy, zz, pl, L_p, M_p, S_p, zcam);
-
-                if (Jz > maxi) {
-                    maxi = Jz;
-                }
-
-                if (Jz < mini) {
-                    mini = Jz;
-                }
-
-                sum += Jz;
-                // I read bz, az values and Hz ==> with low chroma values Hz are very different from lab always around 1.4 radians ???? for blue...
-            }
-        }
-
-        nc = height * width;
-        sum = sum / nc;
-        maxi += epsiljz;
-        sum += epsiljz;
-        //remapping Jz
-        double ijz100 = 1. / jz100;
-        double ajz = (ijz100 - 1.) / 9.; //9 = sqrt(100) - 1 with a parabolic curve after jz100 - we can change for others curve ..log...(you must change also in locallabtool2)
-        double bjz = 1. - ajz;
-        //relation between adapjz and Absolute luminance source (La), adapjz =sqrt(La) - see locallabtool2 adapjzcie
-        double interm = jz100 * (adapjz * ajz + bjz);
-        double bj = (10. - maxi) / 9.;
-        double aj = maxi - bj;
-        double to_screen = (aj * interm + bj) / maxi;
-        //to screen - remapping of Jz in function real scene absolute luminance
-
-        double to_one = 1.;//only for calculation in range 0..1 or 0..32768
-        to_one = 1 / (maxi * to_screen);
-
-        if (adapjz == 10.) { //force original algorithm if La > 10000
-            to_screen = 1.;
-        }
-
-        if (Qtoj) {
-            double xxw = (d50_d65[0][0] * (double) Xw + d50_d65[0][1] * (double) Yw + d50_d65[0][2] * (double) Zw);
-            double yyw = (d50_d65[1][0] * (double) Xw + d50_d65[1][1] * (double) Yw + d50_d65[1][2] * (double) Zw);
-            double zzw = (d50_d65[2][0] * (double) Xw + d50_d65[2][1] * (double) Yw + d50_d65[2][2] * (double) Zw);
-            double L_pa, M_pa, S_pa;
-            Ciecam02::xyz2jzczhz(jzw, azw, bzw, xxw, yyw, zzw, pl, L_pa, M_pa, S_pa, z_cam);
-
-            if (settings->verbose) { //calculate Jz white for use of lightness instead brightness
-                printf("Jzwhite=%f \n", jzw);
-            }
-
-        }
-
-        const std::unique_ptr<LabImage> temp(new LabImage(width, height));
-        const std::unique_ptr<LabImage> tempresid(new LabImage(width, height));
-        const std::unique_ptr<LabImage> tempres(new LabImage(width, height));
-        array2D<double> JJz(width, height);
-        array2D<double> Aaz(width, height);
-        array2D<double> Bbz(width, height);
-        int highhs =  params->locallab.spots.at(sp).hljzcie;
-        int hltonahs = params->locallab.spots.at(sp).hlthjzcie;
-        int shadhs = params->locallab.spots.at(sp).shjzcie;
-        int shtonals = params->locallab.spots.at(sp).shthjzcie;
-        int radhs = params->locallab.spots.at(sp).radjzcie;
-        float softjz = (float) params->locallab.spots.at(sp).softjzcie;
-
-        avgm = 0.5 * (sum * to_screen * to_one + avgm);//empirical formula
-        double miny = 0.1;
-        double delta = 0.015 * (double) sqrt(std::max(100.f, la) / 100.f);//small adaptation in function La scene
-        double maxy = 0.65;//empirical value
-        double maxreal = maxi * to_screen;
-        double maxjzw = jzw * to_screen;
-
-        if (settings->verbose) {
-            printf("La=%4.1f PU_adap=%2.1f maxi=%f mini=%f mean=%f, avgm=%f to_screen=%f Max_real=%f to_one=%f\n", (double) la, adapjz, maxi, mini, sum, avgm, to_screen, maxreal, to_one);
-        }
-        const float sigmoidlambdajz = params->locallab.spots.at(sp).sigmoidldajzcie;
-        const float sigmoidthjz = params->locallab.spots.at(sp).sigmoidthjzcie;
-        const float sigmoidbljz = params->locallab.spots.at(sp).sigmoidbljzcie;
-
-        float thjz = 1.f;
-        const float atjz = 1.f - sigmoidthjz;
-        const float btjz = sigmoidthjz;
-
-        const float athjz = sigmoidthjz - 1.f;
-        const float bthjz = 1.f;
-        float powsig = pow_F(sigmoidlambdajz, 0.5f);
-        const float sigmjz = 3.3f + 7.1f * (1.f - powsig); // e^10.4 = 32860
-        const float bljz = sigmoidbljz;
-
-
-
-        double contreal = 0.2 *  params->locallab.spots.at(sp).contjzcie;
-        DiagonalCurve jz_contrast({
-            DCT_NURBS,
-            0, 0,
-            avgm - avgm * (0.6 - contreal / 250.0), avgm - avgm * (0.6 + contreal / 250.0),
-            avgm + (1. - avgm) * (0.6 - contreal / 250.0), avgm + (1. - avgm) * (0.6 + contreal / 250.0),
-            1, 1
-        });
-        //all calculations in double for best results...but slow
-        double lightreal = 0.2 *  params->locallab.spots.at(sp).lightjzcie;
-        double chromz =  params->locallab.spots.at(sp).chromjzcie;
-        double saturz =  params->locallab.spots.at(sp).saturjzcie;
-        double dhue = 0.0174 * params->locallab.spots.at(sp).huejzcie;
-        DiagonalCurve jz_light({
-            DCT_NURBS,
-            0, 0,
-            miny, miny + lightreal / 150.,
-            maxy, min(1.0, maxy + delta + lightreal / 300.0),
-            1, 1
-        });
-        DiagonalCurve jz_lightn({
-            DCT_NURBS,
-            0, 0,
-            max(0.0, miny  - lightreal / 150.), miny,
-            maxy + delta - lightreal / 300.0, maxy + delta,
-            1, 1
-        });
-        bool wavcurvejz = false;
-
-        if (locwavCurvejz && locwavutilijz) {
-            for (int i = 0; i < 500; i++) {
-                if (locwavCurvejz[i] != 0.5f) {
-                    wavcurvejz = true;
-                    break;
-                }
-            }
-        }
-
-        float mjjz = lp.mLjz;
-
-        if (wavcurvejz && lp.mLjz == 0.f) {
-            mjjz = 0.0f;//to enable clarity if need in some cases mjjz = 0.0001f
-        }
-
-        //log encoding Jz
-        double gray = 0.15;
-        /*
-        const double shadows_range =  params->locallab.spots.at(sp).blackEvjz;
-        const double targetgray = params->locallab.spots.at(sp).targetjz;
-        double targetgraycor = 0.15;
-        double dynamic_range = std::max(params->locallab.spots.at(sp).whiteEvjz - shadows_range, 0.5);
-        const double noise = pow(2., -16.6);//16.6 instead of 16 a little less than others, but we work in double
-        const double log2 = xlog(2.);
-        */
-        double base = 10.;
-        double linbase = 10.;
-
-        if (logjz) { //with brightness Jz
-            gray = 0.01 * params->locallab.spots.at(sp).sourceGraycie;//acts as amplifier (gain) : needs same type of modifications than targetgraycor with pow
-            gray = pow(gray, 1.2);//or 1.15 => modification to increase sensitivity gain, only on defaults, of course we can change this value manually...take into account suuround and Yb Cam16
-            targetgraycor = pow(0.01 * targetgray, 1.15);//or 1.2 small reduce effect -> take into account a part of surround (before it was at 1.2)
-            base = targetgray > 1. && targetgray < 100. && dynamic_range > 0. ? (double) find_gray(std::abs((float) shadows_range) / (float) dynamic_range, (float)(targetgraycor)) : 0.;
-            linbase = std::max(base, 2.);//2. minimal base log to avoid very bad results
-
-            if (settings->verbose) {
-                printf("Base logarithm encoding Jz=%5.1f\n", linbase);
-            }
-        }
-
-        const auto applytojz =
-        [ = ](double x) -> double {
-
-            x = std::max(x, noise);
-            x = std::max(x / gray, noise);//gray = gain - before log conversion
-            x = std::max((xlog(x) / log2 - shadows_range) / dynamic_range, noise);//x in range EV
-            assert(x == x);
-
-            if (linbase > 0.)//apply log base in function of targetgray blackEvjz and Dynamic Range
-            {
-                x = xlog2lin(x, linbase);
-            }
-
-            return x;
-        };
-
-#ifdef _OPENMP
-        #pragma omp parallel for if(multiThread)
-#endif
-
-        for (int i = 0; i < height; i++) {
-            for (int k = 0; k < width; k++) {
-                float L = lab->L[i][k];
-                float a = lab->a[i][k];
-                float b = lab->b[i][k];
-                float x, y, z;
-                //convert Lab => XYZ
-                Color::Lab2XYZ(L, a, b, x, y, z);
-                x = x / 65535.f;
-                y = y / 65535.f;
-                z = z / 65535.f;
-                double Jz, az, bz;//double need because matrix with const(1.6295499532821566e-11) and others
-                double xx, yy, zz;
-                //change WP to D65
-                xx = (d50_d65[0][0] * (double) x + d50_d65[0][1] * (double) y + d50_d65[0][2] * (double) z);
-                yy = (d50_d65[1][0] * (double) x + d50_d65[1][1] * (double) y + d50_d65[1][2] * (double) z);
-                zz = (d50_d65[2][0] * (double) x + d50_d65[2][1] * (double) y + d50_d65[2][2] * (double) z);
-
-                double L_p, M_p, S_p;
-                bool zcam = z_cam;
-                Ciecam02::xyz2jzczhz(Jz, az, bz, xx, yy, zz, pl, L_p, M_p, S_p, zcam);
-                //remapping Jz
-                Jz = Jz * to_screen;
-                az = az * to_screen;
-                bz = bz * to_screen;
-                JJz[i][k] = Jz;
-                Aaz[i][k] = az;
-                Bbz[i][k] = bz;
-
-                if (highhs > 0 || shadhs > 0  || wavcurvejz || mjjz != 0.f || lp.mCjz != 0.f  || LHcurvejz || HHcurvejz || CHcurvejz) {
-                    //here we work in float with usual functions  SH / wavelets / curves H
-                    temp->L[i][k] = tempresid->L[i][k] = tempres->L[i][k] = (float) to_one * 32768.f * (float) JJz[i][k];
-                    temp->a[i][k] = tempresid->a[i][k] = tempres->a[i][k] = (float) to_one * 32768.f * (float) Aaz[i][k];
-                    temp->b[i][k] = tempresid->b[i][k] = tempres->b[i][k] = (float) to_one * 32768.f * (float) Bbz[i][k];
-                }
-            }
-        }
-
-        if (highhs > 0 || shadhs > 0) {
-            ImProcFunctions::shadowsHighlights(temp.get(), true, 1, highhs, shadhs, radhs, sk, hltonahs * maxi * to_screen * to_one, shtonals * maxi * to_screen * to_one);
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int i = 0; i < height; i++) {
-                for (int k = 0; k < width; k++) {//reinitialize datas after SH...: guide, etc.
-                    tempresid->L[i][k] = tempres->L[i][k] = temp->L[i][k];
-                    tempresid->a[i][k] = tempres->a[i][k] = temp->a[i][k];
-                    tempresid->b[i][k] = tempres->b[i][k] = temp->b[i][k];
-                }
-            }
-        }
-
-        //others "Lab" treatment...to adapt
-
-        if (wavcurvejz  || mjjz != 0.f || lp.mCjz != 0.f) { //local contrast wavelet and clarity
-#ifdef _OPENMP
-            const int numThreads = omp_get_max_threads();
-#else
-            const int numThreads = 1;
-
-#endif
-            // adap maximum level wavelet to size of RT-spot
-            int wavelet_level = 1 + params->locallab.spots.at(sp).csthresholdjz.getBottomRight();//retrieve with +1 maximum wavelet_level
-            wavelet_level = rtengine::max(5, wavelet_level);
-            int minwin = rtengine::min(width, height);
-            int maxlevelspot = 9;//maximum possible
-
-            // adapt maximum level wavelet to size of crop
-            while ((1 << maxlevelspot) >= (minwin) && maxlevelspot  > 1) {
-                --maxlevelspot ;
-            }
-
-
-            wavelet_level = rtengine::min(wavelet_level, maxlevelspot);
-            int maxlvl = wavelet_level;
-
-            //simple local contrast in function luminance
-            if (locwavCurvejz && locwavutilijz && wavcurvejz) {
-                float strengthjz = 1.2f;
-                std::unique_ptr<wavelet_decomposition> wdspot(new wavelet_decomposition(temp->L[0], bfw, bfh, maxlvl, 1, sk, numThreads, lp.daubLen));//lp.daubLen
-
-                if (wdspot->memory_allocation_failed()) {
-                    return;
-                }
-
-                maxlvl = wdspot->maxlevel();
-                wavlc(*wdspot, level_bljz, level_hljz, maxlvl, level_hrjz, level_brjz, ahighjz, bhighjz, alowjz, blowjz, sigmalcjz, 1.f, strengthjz, locwavCurvejz, numThreads);
-                wdspot->reconstruct(temp->L[0], 1.f);
-
-            }
-
-            float thr = 0.001f;
-            int flag = 2;
-
-            // begin clarity wavelet jz
-            if (mjjz != 0.f || lp.mCjz != 0.f) {
-                float mL0 = 0.f;
-                float mC0 = 0.f;
-                bool exec = false;
-                float mL = mjjz;
-                float mC = lp.mCjz;
-                clarimerge(lp, mL, mC, exec, tempresid.get(), wavelet_level, sk, numThreads);
-
-                if (maxlvl <= 4) {
-                    mL0 = 0.f;
-                    mC0 = 0.f;
-                    mL = -1.5f * mL;//increase only for sharpen
-                    mC = -mC;
-                    thr = 1.f;
-                    flag = 0;
-
-                } else {
-                    mL0 = mL;
-                    mC0 = mC;
-                    thr = 1.f;
-                    flag = 2;
-                }
-
-                LabImage *mergfile = temp.get();
-#ifdef _OPENMP
-                #pragma omp parallel for if (multiThread)
-#endif
-
-                for (int x = 0; x < height; x++)
-                    for (int y = 0; y < width; y++) {
-                        temp->L[x][y] = clipLoc((1.f + mL0) * mergfile->L[x][y] - mL * tempresid->L[x][y]);
-                        temp->a[x][y] = clipC((1.f + mC0) * mergfile->a[x][y] - mC * tempresid->a[x][y]);
-                        temp->b[x][y] = clipC((1.f + mC0) * mergfile->b[x][y] - mC * tempresid->b[x][y]);
-                    }
-            }
-
-            if (lp.softrjz >= 0.5f && (wavcurvejz || std::fabs(mjjz) > 0.001f)) {//guidedfilter
-                softproc(tempres.get(), temp.get(), lp.softrjz, height, width, 0.001, 0.00001, thr, sk, multiThread, flag);
-            }
-        }
-
-//new curves Hz
-#ifdef _OPENMP
-        #pragma omp parallel for if (multiThread)
-#endif
-
-        for (int i = 0; i < height; i++) {
-            for (int k = 0; k < width; k++) {
-                float j_z = temp->L[i][k];
-                float C_z = sqrt(SQR(temp->a[i][k]) + SQR(temp->b[i][k]));
-                float c_z = C_z / 32768.f;
-
-                if (loclhCurvejz && LHcurvejz) {//Jz=f(Hz) curve
-                    float kcz = (float) jzamountchr;
-                    float Hz = xatan2f(temp->b[i][k], temp->a[i][k]);
-                    float l_r = j_z / 32768.f;
-                    float kcc = SQR(c_z / kcz);
-                    jzch = true;
-
-                    if (jzch == false) {
-                        kcc = 1.f;
-                    } else if (kcc > 1.f) {
-                        kcc = 1.f; //cbrt(kcc);
-                    }
-
-                    float valparam = loclhCurvejz[500.f * static_cast<float>(Color::huejz_to_huehsv2((float) Hz))] - 0.5f;
-
-                    float valparamneg;
-                    valparamneg = valparam;
-                    valparam *= 2.f * kcc;
-                    valparamneg *= kcc;
-
-                    if (valparam > 0.f) {
-                        l_r = (1.f - valparam) * l_r + valparam * (1.f - SQR(((SQR(1.f - min(l_r, 1.0f))))));
-                    } else
-                        //for negative
-                    {
-                        float khue = 1.9f; //in reserve in case of!
-                        l_r *= (1.f + khue * valparamneg);
-                    }
-
-                    temp->L[i][k] = l_r * 32768.f;
-                }
-
-                if (locchCurvejz && CHcurvejz) {//Cz=f(Hz) curve
-                    float Hz = xatan2f(temp->b[i][k], temp->a[i][k]);
-                    const float valparam = 1.5f * (locchCurvejz[500.f * static_cast<float>(Color::huejz_to_huehsv2((float)Hz))] - 0.5f);  //get valp=f(H)
-                    float chromaCzfactor = 1.0f + valparam;
-                    temp->a[i][k] *= chromaCzfactor;
-                    temp->b[i][k] *= chromaCzfactor;
-                }
-
-
-                if (lochhCurvejz && HHcurvejz) { // Hz=f(Hz)
-                    float Hz = xatan2f(temp->b[i][k], temp->a[i][k]);
-                    const float valparam = 1.4f * (lochhCurvejz[500.f * static_cast<float>(Color::huejz_to_huehsv2((float)Hz))] - 0.5f) + static_cast<float>(Hz);
-                    Hz = valparam;
-
-                    if (Hz < 0.0f) {
-                        Hz += (2.f * rtengine::RT_PI_F);
-                    }
-
-                    float2 sincosval = xsincosf(Hz);
-                    temp->a[i][k] = C_z * sincosval.y;
-                    temp->b[i][k] = C_z * sincosval.x;
-                }
-            }
-        }
-
-        if (loclhCurvejz && LHcurvejz && softjz > 0.f) {//Guidedilter for artifacts curve J(H)
-            float thr = 0.00001f;
-            int flag = 2;
-            float softjzr = 0.05f * softjz;
-            softproc(tempres.get(), temp.get(), softjzr, height, width, 0.000001, 0.00000001, thr, sk, multiThread, flag);
-        }
-
-
-        if ((lochhCurvejz && HHcurvejz) || (locchCurvejz && CHcurvejz)) { //for artifacts curve H(H)
-            if (softjz > 0.f) {
-                array2D<float> chro(width, height);
-                array2D<float> hue(width, height);
-                array2D<float> guid(width, height);
-
-#ifdef _OPENMP
-                #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-                for (int y = 0; y < height; y++) {
-                    for (int x = 0; x < width; x++) {
-                        hue[y][x] = xatan2f(temp->b[y][x], temp->a[y][x]);
-                        chro[y][x] = sqrt(SQR(temp->b[y][x]) + SQR(temp->a[y][x])) / 32768.f;
-
-                        if (hue[y][x] < 0.0f) {
-                            hue[y][x] += (2.f * rtengine::RT_PI_F);
-                        }
-
-                        hue[y][x] /= (2.f * rtengine::RT_PI_F);
-                        guid[y][x] = tempres->L[y][x] / 32768.f;
-                    }
-                }
-
-                float softr = softjz;
-                const float tmpblur = softr < 0.f ? -1.f / softr : 1.f + softr;
-                const int r2 = rtengine::max<int>(10 / sk * tmpblur + 0.2f, 1);
-                const int r1 = rtengine::max<int>(4 / sk * tmpblur + 0.5f, 1);
-                constexpr float epsilmax = 0.0005f;
-                constexpr float epsilmin = 0.0000001f;
-                constexpr float aepsil = (epsilmax - epsilmin) / 100.f;
-                constexpr float bepsil = epsilmin;
-                const float epsil = softr < 0.f ? 0.001f : aepsil * softr + bepsil;
-
-                if (lochhCurvejz && HHcurvejz) {
-                    rtengine::guidedFilter(guid, hue, hue, r2, 0.5f * epsil, multiThread);
-                }
-
-                if (locchCurvejz && CHcurvejz) {
-                    rtengine::guidedFilter(guid, chro, chro, r1, 0.4f * epsil, multiThread);
-                }
-
-#ifdef _OPENMP
-                #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-                for (int y = 0; y < height; y++) {
-                    for (int x = 0; x < width; x++) {
-                        hue[y][x] *= (2.f * rtengine::RT_PI_F);
-                        chro[y][x] *= 32768.f;
-                        float2 sincosval = xsincosf(hue[y][x]);
-                        temp->a[y][x] = chro[y][x] * sincosval.y;
-                        temp->b[y][x] = chro[y][x] * sincosval.x;
-                    }
-                }
-            }
-        }
-
-
-///////////////////
-
-
-#ifdef _OPENMP
-        #pragma omp parallel  for if(multiThread)
-#endif
-
-        for (int i = 0; i < height; i++) {
-            for (int k = 0; k < width; k++) {
-                //reconvert to double
-                if (highhs > 0 || shadhs > 0  || wavcurvejz || mjjz != 0.f || lp.mCjz != 0.f || LHcurvejz || HHcurvejz || CHcurvejz) {
-                    //now we work in double necessary for matrix conversion and when in range 0..1 with use of PQ
-                    JJz[i][k] = (double)(temp->L[i][k] / (32768.f * (float) to_one));
-                    Aaz[i][k] = (double)(temp->a[i][k] / (32768.f * (float) to_one));
-                    Bbz[i][k] = (double)(temp->b[i][k] / (32768.f * (float) to_one));
-                }
-
-                double az =  Aaz[i][k];
-                double bz =  Bbz[i][k];
-                double Jz =  LIM01(JJz[i][k]);
-                Jz *= to_one;
-                double Cz = sqrt(az * az + bz * bz);
-
-                //log encoding
-                if (logjz) {
-                    double jmz =  Jz;
-
-                    if (jmz > noise) {
-                        double mm = applytojz(jmz);
-                        double f = mm / jmz;
-                        Jz *= f;
-                        Jz = LIM01(Jz);//clip values
-                    }
-                }
-
-                //sigmoid 5.12
-                if (issigjz12 && iscie && modeqj == 1) { //sigmoid Jz
-                    float val = Jz;
-                    float Jout = 0.f;
-                    sigmoid_QJ(val, Jout, middle_grey_contrastjz, contrast_skewnessjz, middle_greyjz, black_pointjz, white_point_dispjz);
-
-                    Jz = Jout;
-                    Jz = LIM01(Jz);
-                }
-
-                //sigmoid 5.11
-                if (issigjz && iscie && modeqj == 0) { //sigmoid Jz
-                    float val = Jz;
-
-                    if (islogjz) {
-                        val = std::max((xlog(Jz) / log2 - shadows_range) / (dynamic_range + 1.5), noise);//in range EV
-                    }
-
-                    if (sigmoidthjz >= 1.f) {
-                        thjz = athjz * val + bthjz;//threshold
-                    } else {
-                        thjz = atjz * val + btjz;
-                    }
-
-                    sigmoidla(val, thjz, sigmjz); //sigmz "slope" of sigmoid
-
-
-                    Jz = LIM01((double) bljz * Jz + (double) val);
-                }
-
-                if (Qtoj == true) { //lightness instead of brightness
-                    Jz /= to_one;
-                    Jz /= maxjzw;//Jz white
-                    Jz = SQR(Jz);
-                }
-
-                //contrast
-                Jz = LIM01(jz_contrast.getVal(LIM01(Jz)));
-
-                //brightness and lightness
-                if (lightreal > 0) {
-                    Jz = LIM01(jz_light.getVal(Jz));
-                }
-
-                if (lightreal < 0) {
-                    Jz = LIM01(jz_lightn.getVal(Jz));
-                }
-
-                //Jz (Jz) curve
-                double Jzold = Jz;
-
-                if (jzlocalcurve && localjzutili) {
-                    Jz = (double)(jzlocalcurve[(float) Jz * 65535.f] / 65535.f);
-                    Jz  = 0.3 * (Jz - Jzold) + Jzold;
-                }
-
-                //reconvert from lightness or Brightness
-                if (Qtoj == false) {
-                    Jz /= to_one;
-                } else {
-                    Jz = sqrt(Jz);
-                    Jz *= maxjzw;
-                }
-
-                double Hz;
-                //remapping Cz
-                Hz = xatan2(bz, az);
-                double Czold = Cz;
-
-                //Cz(Cz) curve
-                if (czlocalcurve && localczutili) {
-                    Cz = (double)(czlocalcurve[(float) Cz * 92666.f * (float) to_one] / (92666.f * (float) to_one));
-                    Cz  = 0.5 * (Cz - Czold) + Czold;
-                }
-
-                //Cz(Jz) curve
-                if (czjzlocalcurve && localczjzutili) {
-                    double chromaCfactor = (double)(czjzlocalcurve[(float) Jz * 65535.f * (float) to_one]) / (Jz * 65535. * to_one);
-                    Cz  *=  chromaCfactor;
-                }
-
-                //Hz in 0 2*PI
-                if (Hz < 0.0) {
-                    Hz += (2. * rtengine::RT_PI);
-                }
-
-                //Chroma slider
-                if (chromz < 0.) {
-                    Cz = Cz * (1. + 0.01 * chromz);
-                } else {
-                    double maxcz = czlim / to_one;
-                    double fcz = Cz / maxcz;
-                    double pocz = pow(fcz, 1. - 0.0024 * chromz); //increase value - before 0.0017
-                    Cz = maxcz * pocz;
-                    //  Cz = Cz * (1. + 0.005 * chromz);//linear
-                }
-
-                //saturation slider
-                if (saturz != 0.) {
-                    double js = Jz / maxjzw; //divide by Jz white
-                    js = SQR(js);
-
-                    if (js <= 0.) {
-                        js = 0.0000001;
-                    }
-
-                    double Sz = Cz / (js);
-
-                    if (saturz < 0.) {
-                        Sz = Sz * (1. + 0.01 * saturz);
-                    } else {
-                        Sz = Sz * (1. + 0.003 * saturz);//not pow function because Sz is "open" - 0.003 empirical value to have results comparable to Cz
-                    }
-
-                    Cz = Sz * js;
-                }
-
-                //rotation hue
-                Hz += dhue;
-
-                if (Hz < 0.0) {
-                    Hz += (2. * rtengine::RT_PI);
-                }
-
-                Cz = clipcz(Cz);
-                double2 sincosval = xsincos(Hz);
-                az = clipazbz(Cz * sincosval.y);
-                bz = clipazbz(Cz * sincosval.x);
-                Cz = sqrt(az * az + bz * bz);
-
-
-                bz = bz / (to_screen);
-                az = az / (to_screen);
-
-                Jz = LIM01(Jz / (to_screen));
-
-                if (jabcie) { //Not used does not work at all
-                    Jz = clipjz05(Jz);
-                    gamutjz(Jz, az, bz, pl, wip, 0.94, 0.004);
-                }
-
-                double L_, M_, S_;
-                double xx, yy, zz;
-                bool zcam = z_cam;
-                //reconvert to XYZ in double
-                Ciecam02::jzczhzxyz(xx, yy, zz, Jz, az, bz, pl, L_, M_, S_, zcam);
-                //re enable D50
-                double x, y, z;
-                x = 65535. * (d65_d50[0][0] * xx + d65_d50[0][1] * yy + d65_d50[0][2] * zz);
-                y = 65535. * (d65_d50[1][0] * xx + d65_d50[1][1] * yy + d65_d50[1][2] * zz);
-                z = 65535. * (d65_d50[2][0] * xx + d65_d50[2][1] * yy + d65_d50[2][2] * zz);
-
-                float Ll, aa, bb;
-                Color::XYZ2Lab(x,  y,  z, Ll, aa, bb);
-                lab->L[i][k] = Ll;
-                lab->a[i][k] = aa;
-                lab->b[i][k] = bb;
-            }
-        }
-    }
-                    //lp.sursouci==4 disable ciecam
-    if ((mocam == 1 && lp.sursouci!= 4)|| mocam ==2 || call == 1  || call == 2 || call == 10) { //CAM16 call=2 vibrance warm-cool - call = 10 take into account "mean luminance Yb for Jz
-//begin ciecam
-        if (settings->verbose && (mocam == 1  || call == 1)) {//display only if choice cam16
-            //information on Cam16 scene conditions - allows user to see choices's incidences
-            float maxicamq = -1000.f;
-            float maxicamj = -1000.f;
-            float maxisat = -1000.f;
-            float maxiM = -1000.f;
-            float minicam = 1000000.f;
-            float minicamq = 1000000.f;
-            float minisat = 1000000.f;
-            float miniM = 1000000.f;
-            int nccam = 0;
-            float sumcam = 0.f;
-            float sumcamq = 0.f;
-            float sumsat = 0.f;
-            float sumM = 0.f;
-
-            if (lp.logena && !(params->locallab.spots.at(sp).expcie && mocam == 1)) { //Log encoding only, but enable for log encoding if we use Cam16 module both with log encoding
-                plum = 100.f;
-            }
-
-            //find main values Cam16
-#ifdef _OPENMP
-            #pragma omp parallel for reduction(min:minicam) reduction(max:maxicamj) reduction(min:minicamq) reduction(max:maxicamq) reduction(min:minisat) reduction(max:maxisat) reduction(min:miniM) reduction(max:maxiM) reduction(+:sumcam) reduction(+:sumcamq) reduction(+:sumsat) reduction(+:sumM)if(multiThread)
-#endif
-
-            for (int i = 0; i < height; i += 1) {
-                for (int k = 0; k < width; k += 1) {
-                    float L = lab->L[i][k];
-                    float a = lab->a[i][k];
-                    float b = lab->b[i][k];
-                    float x, y, z;
-                    //convert Lab => XYZ
-                    Color::Lab2XYZ(L, a, b, x, y, z);
-                    x = x / 655.35f;
-                    y = y / 655.35f;
-                    z = z / 655.35f;
-                    float J, C, h, Q, M, s;
-                    Ciecam02::xyz2jchqms_ciecam02float(J, C,  h,
-                                                       Q,  M,  s, aw, fl, wh,
-                                                       x,  y,  z,
-                                                       xw1, yw1,  zw1,
-                                                       c,  nc, pow1, nbb, ncb, pfl, cz, d, c16, plum);
-
-                    if (J > maxicamj) {
-                        maxicamj = J;
-                    }
-
-                    if (J < minicam) {
-                        minicam = J;
-                    }
-
-                    sumcam += J;
-
-                    if (Q > maxicamq) {
-                        maxicamq = Q;
-                    }
-
-                    if (Q < minicamq) {
-                        minicamq = Q;
-                    }
-
-                    sumcamq += Q;
-
-                    if (s > maxisat) {
-                        maxisat = s;
-                    }
-
-                    if (s < minisat) {
-                        minisat = s;
-                    }
-
-                    sumsat += s;
-
-                    if (M > maxiM) {
-                        maxiM = M;
-                    }
-
-                    if (M < miniM) {
-                        miniM = M;
-                    }
-
-                    sumM += M;
-                }
-            }
-
-            nccam = height * width;
-            sumcam = sumcam / nccam;
-            sumcamq /= nccam;
-            sumsat /= nccam;
-            sumM /= nccam;
-            
-            if (settings->verbose) {
-                printf("Cam16 Scene  Lighness_J Brightness_Q- HDR-PQ=%5.1f minJ=%3.1f maxJ=%3.1f meanJ=%3.1f minQ=%3.1f maxQ=%4.1f  meanQ=%4.1f meanQ1=%2.3f\n", (double) plum, (double) minicam, (double) maxicamj, (double) sumcam, (double) minicamq, (double) maxicamq, (double) sumcamq, (double) (sumcamq * coefq));
-                printf("Cam16 Scene  Saturati-s Colorfulln_M- minSat=%3.1f maxSat=%3.1f meanSat=%3.1f minM=%3.1f maxM=%3.1f meanM=%3.1f\n", (double) minisat, (double) maxisat, (double) sumsat, (double) miniM, (double) maxiM, (double) sumM);
-            }
-           // maxicam = maxicamq;//maximum Brightness 
-            if(sumcamq < maxicamq) {
-                // maxicam = sumcamq + 0.2f * minicamq;//maximum Brightness take into account
-                 maxicam = sumcamq;//maximum Brightness take into account
-                //ponderate maxicam with mean and mini
-            } else {
-                maxicam = 0.4f * sumcamq + 0.6f * maxicamq;               
-            } 
-            sumcamq01 = sumcamq * coefq;
-
-        }
-
-        float base = 10.f;
-        float linbaseor = 10.f;
-        float linbase = 10.f;
-        float gray = 15.f;
-
-        const bool compr = params->locallab.spots.at(sp).comprcie > 0.;
-        float comprfactor = params->locallab.spots.at(sp).comprcie;
-        float comprth = 1.f; //0.1 +  params->locallab.spots.at(sp).comprcieth;
-
-        double drref = 8.5; //Dynamic Range standard
- 
-        double drd = ((double) dynamic_range - drref) / drref;
- 
-        double dratt = (double) dynamic_range / drref;
-        comprfactor = 0.4f * comprfactor * (float) dratt;//adapt comprfactor to Dynamic Range
-        float newgray = 0.18f;
-
-
-     //   bool logqprov = false; 
-        if ((params->locallab.spots.at(sp).logcie && params->locallab.spots.at(sp).logcieq) || mobwev != 0) {//increase Dyn Range when log encoding
-            dynamic_range += 0.2;//empirical value
-            gray = 0.01f * (float) params->locallab.spots.at(sp).sourceGraycie;
-            const float targetgraycie = params->locallab.spots.at(sp).targetGraycie;
-            float targetgraycor = 0.01f * targetgraycie;
-            base = targetgraycie > 1.f && targetgraycie < 100.f && (float) dynamic_range > 0.f ?  find_gray(std::abs((float) shadows_range) / (float) dynamic_range, (targetgraycor)) : 0.f;
-            linbaseor = std::max(base, 2.f);//2. minimal base log to avoid very bad results
-
-            float maxQgray = coefq * maxicam / gray;
-            maxicam =  maxQgray;//setting threshold comprcieth
-            const float log2 = xlogf(2.f);
-
-            float corlog = xlogf(maxicam)/log2;//correction base logarithme
-            linbase = linbaseor / corlog;
-            newgray = gray; //gray - 0.022f * (6.f - maxicam);//empirical formula to take into account Q in DR. 6.f  =>approach to mean overall images
-
-            if (settings->verbose) {
-                printf("Gray=%1.3f newgray=%1.3f MaxicamQ=%3.2f Base log encode corrected Q=%5.1f Base log encode origig Q=%5.1f\n", (double) gray, (double) newgray, (double) maxicam, (double) linbase, (double) linbaseor);
-            }
-
-        }
-
-        const auto applytoq =
-        [ = ](float x) -> float {
-
-            x = rtengine::max(x, (float) noise);
-            x = rtengine::max(x / newgray, (float) noise);//gray = gain - before log conversion
-
-            if (compr && x >= comprth)//comprth = maxicam
-            {
-                x = intp(comprfactor, (std::tanh((x - comprth) / comprth) + 1.f) * comprth, x); //as sigmoid... but tanh (tg hyperbolic), inspired by the work of alberto Grigio
-            }
-        
-            x = rtengine::max((xlogf(x) / log2f - (float) shadows_range) / (float) dynamic_range, (float) noise);//x in range EV
-            assert(x == x);
-
-            if (linbase > 0.f)//apply log base in function of targetgray blackEvjz and Dynamic Range
-            {
-                x = xlog2lin(x, linbase);
-            }
-
-            return x;
-        };
-
-        //prepare Normalize luminance
-        float *datain = nullptr;
-        float *data = nullptr;
-        float *datanorm = nullptr;
-
-        if (((sigmoidnorm  && issigq)  || params->locallab.spots.at(sp).logcieq) && modeqj == 0) {//5.11
-            datain = new float[width* height];
-            data = new float[width * height];
-            datanorm = new float[width * height];
-#ifdef _OPENMP
-            #pragma omp parallel for schedule(dynamic, 16)
-#endif
-
-            for (int y = 0; y < height; y++) {
-                for (int x = 0; x < width; x++) {
-                    datain[(y) * width + (x)] = lab->L[y][x];
-                }
-            }
-        }
-
-#if defined(__SSE2__) || defined(RT_SIMDE)
-        int bufferLength = ((width + 3) / 4) * 4; // bufferLength has to be a multiple of 4
-#endif
-#ifdef _OPENMP
-        #pragma omp parallel if (multiThread)
-#endif
-        {
-#if defined(__SSE2__) || defined(RT_SIMDE)
-            // one line buffer per channel and thread
-            float Jbuffer[bufferLength] ALIGNED16;
-            float Cbuffer[bufferLength] ALIGNED16;
-            float hbuffer[bufferLength] ALIGNED16;
-            float Qbuffer[bufferLength] ALIGNED16;
-            float Mbuffer[bufferLength] ALIGNED16;
-            float sbuffer[bufferLength] ALIGNED16;
-#endif
-#ifdef _OPENMP
-            #pragma omp for schedule(dynamic, 16)
-#endif
-
-            for (int i = 0; i < height; i++) {
-#if defined(__SSE2__) || defined(RT_SIMDE)
-                // vectorized conversion from Lab to jchqms
-                int k;
-                vfloat c655d35 = F2V(655.35f);
-
-                for (k = 0; k < width - 3; k += 4) {
-                    vfloat x, y, z;
-                    Color::Lab2XYZ(LVFU(lab->L[i][k]), LVFU(lab->a[i][k]), LVFU(lab->b[i][k]), x, y, z);
-                    x = x / c655d35;
-                    y = y / c655d35;
-                    z = z / c655d35;
-                    vfloat J, C, h, Q, M, s;
-                    Ciecam02::xyz2jchqms_ciecam02float(J, C,  h,
-                                                       Q,  M,  s, F2V(aw), F2V(fl), F2V(wh),
-                                                       x,  y,  z,
-                                                       F2V(xw1), F2V(yw1),  F2V(zw1),
-                                                       F2V(c),  F2V(nc), F2V(pow1), F2V(nbb), F2V(ncb), F2V(pfl), F2V(cz), F2V(d), c16, F2V(plum));
-                    STVF(Jbuffer[k], J);
-                    STVF(Cbuffer[k], C);
-                    STVF(hbuffer[k], h);
-                    STVF(Qbuffer[k], Q);
-                    STVF(Mbuffer[k], M);
-                    STVF(sbuffer[k], s);
-                }
-
-                for (; k < width; k++) {
-                    float L = lab->L[i][k];
-                    float a = lab->a[i][k];
-                    float b = lab->b[i][k];
-                    float x, y, z;
-                    //convert Lab => XYZ
-                    Color::Lab2XYZ(L, a, b, x, y, z);
-                    x = x / 655.35f;
-                    y = y / 655.35f;
-                    z = z / 655.35f;
-                    float J, C, h, Q, M, s;
-                    Ciecam02::xyz2jchqms_ciecam02float(J, C,  h,
-                                                       Q,  M,  s, aw, fl, wh,
-                                                       x,  y,  z,
-                                                       xw1, yw1,  zw1,
-                                                       c,  nc, pow1, nbb, ncb, pfl, cz, d, c16, plum);
-                    Jbuffer[k] = J;
-                    Cbuffer[k] = C;
-                    hbuffer[k] = h;
-                    Qbuffer[k] = Q;
-                    Mbuffer[k] = M;
-                    sbuffer[k] = s;
-                }
-
-#endif // __SSE2__
-
-                for (int j = 0; j < width; j++) {
-                    float J, C, h, Q, M, s;
-
-#if defined(__SSE2__) || defined(RT_SIMDE)
-                    // use precomputed values from above
-                    J = Jbuffer[j];
-                    C = Cbuffer[j];
-                    h = hbuffer[j];
-                    Q = Qbuffer[j];
-                    M = Mbuffer[j];
-                    s = sbuffer[j];
-#else
-                    float x, y, z;
-                    float L = lab->L[i][j];
-                    float a = lab->a[i][j];
-                    float b = lab->b[i][j];
-                    float x1, y1, z1;
-                    //convert Lab => XYZ
-                    Color::Lab2XYZ(L, a, b, x1, y1, z1);
-                    x = x1 / 655.35f;
-                    y = y1 / 655.35f;
-                    z = z1 / 655.35f;
-                    //process source==> normal
-                    Ciecam02::xyz2jchqms_ciecam02float(J, C,  h,
-                                                       Q,  M,  s, aw, fl, wh,
-                                                       x,  y,  z,
-                                                       xw1, yw1,  zw1,
-                                                       c,  nc, pow1, nbb, ncb, pfl, cz, d, c16, plum);
-#endif
-                    float Jpro, Cpro, hpro, Qpro, Mpro, spro;
-                    Jpro = J;
-                    Cpro = C;
-                    hpro = h;
-                    Qpro = Q;
-                    Mpro = M;
-                    spro = s;
-
-                    if (ciec  && mocam == 1) {//only Cam16
-                        bool jp = false;
-
-                      //  if (params->locallab.spots.at(sp).logcie && iscie) {//log encoding Q
-                        if (params->locallab.spots.at(sp).logcie && params->locallab.spots.at(sp).logcieq && iscie  && modeqj == 0) {//log encoding Q 5.11
-             
-                            float val =  Qpro *  coefq;
-
-                            if (val > (float) noise) {
-                                float mm = applytoq(val);
-                                float f = mm / val;
-                                Qpro *=  f;
-                            }
-                        }
-                        if (issig && issigq12 && iscie && modeqj == 1) { //sigmoid Q and slope based Q 5.12
-                            float val = Qpro * coefq;
-                            float Qout = 0.f;
-                            if(mobwev12 == 0) {
-                                sigmoid_QJ(val, Qout, middle_grey_contrast, contrast_skewness, middle_grey, black_point, white_point_disp);
-                            }
-                            if(mobwev12 == 1) {
-                                bool rolloff = false;//all range
-                                bool kmid = false;//not take into account Yb viewing
-                                tonemapFreemanQ(val, Qout, slopsmootq , white_pointsig, black_point, middle_grey, mid_gray_view, rolloff, kmid);
-                            }
-
-                            Qpro = std::max(Qout / coefq, 0.f);
-                            Jpro = SQR((10.f * Qpro) / wh);
-
-                        }
-
-                        if (issig && issigq && iscie && mobwev != 2 && modeqj == 0) { //sigmoid Q only and black Ev & white Ev 5.11
-
-                           float val = Qpro * coefq;
-
-                            if (mobwev == 1) {
-                                val = std::max((xlog(val) / log2 - shadows_range) / (dynamic_range + 1.5), noise);//in range EV
-                            }
-
-                            float sigreal = sigmoidth * sumcamq01;//correction for sigmoid Q take into account mean Q
-
-                            if (sigreal >= 1.f) {
-                                th = (sigreal - 1.f) * val + 1.f;
-                            } else {
-                                th = (1.f - sigreal) * val + sigreal;
-                            }
-
-                            sigmoidla(val, th, sigm);
-                            Qpro = std::max(Qpro + val / coefq, 0.f);
-                            Qpro = CAMBrightCurveQsig[(float)(Qpro * coefQ)] / coefQ;   //brightness and contrast
-                            
-                            Jpro = SQR((10.f * Qpro) / wh);
-
-                        }
-
-                        if ((cielocalcurve && localcieutili) && mecamcurve == 1) {//curve Q
-                            jp = true;
-                            float Qq = Qpro * coefQ;
-                            float Qold = Qpro;
-                            Qq = 0.5f * cielocalcurve[Qq * 2.f];
-                            Qq = Qq / coefQ;
-                            Qpro = 0.2f * (Qq - Qold) + Qold;
-
-                            if (jp) {
-                                Jpro = SQR((10.f * Qpro) / wh);
-                            }
-                        }
-
-                        Qpro = CAMBrightCurveQ[(float)(Qpro * coefQ)] / coefQ;   //brightness and contrast
-
-                        float Mp, sres;
-                        Mp = Mpro / 100.0f;
-                        Ciecam02::curvecolorfloat(mchr, Mp, sres, 2.5f);
-                        float dred = 100.f; //in C mode
-                        float protect_red = 80.0f; // in C mode
-                        dred *= coe; //in M mode
-                        protect_red *= coe; //M mode
-                        Color::skinredfloat(Jpro, hpro, sres, Mp, dred, protect_red, 0, rstprotection, 100.f, Mpro);
-                        Jpro = SQR((10.f * Qpro) / wh);
-                        Qpro = (Qpro == 0.f ? epsil : Qpro); // avoid division by zero
-                        spro = 100.0f * sqrtf(Mpro / Qpro);
-                        Jpro = CAMBrightCurveJ[(float)(Jpro * 327.68f)];   //lightness CIECAM02 + contrast
-                        float Sp = spro / 100.0f;
-                        Ciecam02::curvecolorfloat(schr, Sp, sres, 1.5f);
-                        dred = 100.f; // in C mode
-                        protect_red = 80.0f; // in C mode
-                        dred = 100.0f * sqrtf((dred * coe) / Q);
-                        protect_red = 100.0f * sqrtf((protect_red * coe) / Q);
-                        Color::skinredfloat(Jpro, hpro, sres, Sp, dred, protect_red, 0, rstprotection, 100.f, spro);
-                        Qpro = QproFactor * sqrtf(Jpro);
-                        float Cp = (spro * spro * Qpro) / (1000000.f);
-                        Cpro = Cp * 100.f;
-                        Ciecam02::curvecolorfloat(cchr, Cp, sres, 1.8f);
-                        Color::skinredfloat(Jpro, hpro, sres, Cp, 55.f, 30.f, 1, rstprotection, 100.f, Cpro);
-
-                        hpro = hpro + hue;
-
-                        if (hpro < 0.0f) {
-                            hpro += 360.0f;    //hue
-                        }
-
-                        if ((cielocalcurve && localcieutili) && mecamcurve == 0) {//curve J
-                            float Jj = (float) Jpro * 327.68f;
-                            float Jold = Jj;
-                            Jj = 0.5f * cielocalcurve[Jj * 2.f];
-                            Jj = 0.3f * (Jj - Jold) + Jold;    //divide sensibility
-                            Jpro = (float)(Jj / 327.68f);
-
-                            if (Jpro < 1.f) {
-                                Jpro = 1.f;
-                            }
-                        }
-
-                        if (cielocalcurve2 && localcieutili2) {//chroma saturation colorfullness
-                            if (mecamcurve2 == 0) {//chroma
-                                float parsat = 0.8f; //0.68;
-                                float coef = 327.68f / parsat;
-                                float Cc = (float) Cpro * coef;
-                                float Ccold = Cc;
-                                Cc = 0.5f * cielocalcurve2[Cc * 2.f];
-                                float dred = 55.f;
-                                float protect_red = 30.0f;
-                                int sk1 = 1;
-                                float ko = 1.f / coef;
-                                Color::skinredfloat(Jpro, hpro, Cc, Ccold, dred, protect_red, sk1, rstprotection, ko, Cpro);
-                            } else if (mecamcurve2 == 1) {//saturation
-                                float parsat = 0.8f; //0.6
-                                float coef = 327.68f / parsat;
-                                float Ss = (float) spro * coef;
-                                float Sold = Ss;
-                                Ss = 0.5f * cielocalcurve2[Ss * 2.f];
-                                Ss = 0.6f * (Ss - Sold) + Sold; //divide sensibility saturation
-                                float dred = 100.f; // in C mode
-                                float protect_red = 80.0f; // in C mode
-                                dred = 100.0f * sqrtf((dred * coe) / Qpro);
-                                protect_red = 100.0f * sqrtf((protect_red * coe) / Qpro);
-                                float ko = 1.f / coef;
-                                Color::skinredfloat(Jpro, hpro, Ss, Sold, dred, protect_red, 0, rstprotection, ko, spro);
-                                Qpro = (4.0f / c) * sqrtf(Jpro / 100.0f) * (aw + 4.0f) ;
-                                Cpro = (spro * spro * Qpro) / (10000.0f);
-                            } else if (mecamcurve2 == 2) {//colorfullness
-                                float parsat = 0.8f; //0.68;
-                                float coef = 327.68f / parsat;
-                                float Mm = (float) Mpro * coef;
-                                float Mold = Mm;
-                                Mm = 0.5f * cielocalcurve2[Mm * 2.f];
-                                float dred = 100.f; //in C mode
-                                float protect_red = 80.0f; // in C mode
-                                dred *= coe; //in M mode
-                                protect_red *= coe;
-                                float ko = 1.f / coef;
-                                Color::skinredfloat(Jpro, hpro, Mm, Mold, dred, protect_red, 0, rstprotection, ko, Mpro);
-                                Cpro = Mpro / coe;
-                            }
-                        }
-                    }
-
-                    //retrieve values C,J...s
-                    C = Cpro;
-                    J = Jpro;
-                    Q = Qpro;
-                    M = Mpro;
-                    h = hpro;
-                    s = spro;
-
-#if defined(__SSE2__) || defined(RT_SIMDE)
-                    // write to line buffers
-                    Jbuffer[j] = J;
-                    Cbuffer[j] = C;
-                    hbuffer[j] = h;
-#else
-                    float xx, yy, zz;
-                    //process normal==> viewing
-
-                    Ciecam02::jch2xyz_ciecam02float(xx, yy, zz,
-                                                    J,  C, h,
-                                                    xw2, yw2,  zw2,
-                                                    c2, nc2,  pow1n, nbbj, ncbj, flj, czj, dj, awj, c16, plum);
-                    x = CLIP(xx * 655.35f);
-                    y = CLIP(yy * 655.35f);
-                    z = CLIP(zz * 655.35f);
-                    float Ll, aa, bb;
-                    //convert xyz=>lab
-                    Color::XYZ2Lab(x,  y,  z, Ll, aa, bb);
-                    lab->L[i][j] = Ll;
-                    lab->a[i][j] = aa;
-                    lab->b[i][j] = bb;
-#endif
-                }
-
-#if defined(__SSE2__) || defined(RT_SIMDE)
-                // process line buffers
-                float *xbuffer = Qbuffer;
-                float *ybuffer = Mbuffer;
-                float *zbuffer = sbuffer;
-
-                for (k = 0; k < bufferLength; k += 4) {
-                    vfloat x, y, z;
-                    Ciecam02::jch2xyz_ciecam02float(x, y, z,
-                                                    LVF(Jbuffer[k]), LVF(Cbuffer[k]), LVF(hbuffer[k]),
-                                                    F2V(xw2), F2V(yw2), F2V(zw2),
-                                                    F2V(nc2), F2V(pow1n), F2V(nbbj), F2V(ncbj), F2V(flj), F2V(dj), F2V(awj), F2V(reccmcz), c16, F2V(plum));
-                    STVF(xbuffer[k], x * c655d35);
-                    STVF(ybuffer[k], y * c655d35);
-                    STVF(zbuffer[k], z * c655d35);
-                }
-
-                // XYZ2Lab uses a lookup table. The function behind that lut is a cube root.
-                // SSE can't beat the speed of that lut, so it doesn't make sense to use SSE
-                for (int j = 0; j < width; j++) {
-                    float Ll, aa, bb;
-                    //convert xyz=>lab
-                    xbuffer[j] = CLIP(xbuffer[j]);
-                    ybuffer[j] = CLIP(ybuffer[j]);
-                    zbuffer[j] = CLIP(zbuffer[j]);
-
-                    Color::XYZ2Lab(xbuffer[j], ybuffer[j], zbuffer[j], Ll, aa, bb);
-
-                    lab->L[i][j] = Ll;
-                    lab->a[i][j] = aa;
-                    lab->b[i][j] = bb;
-                }
-
-#endif
-            }
-        }
-         if (((mocam == 1 && (sigmoidnorm && issigq)) || params->locallab.spots.at(sp).logcieq) && modeqj == 0) { //Normalize luminance 5.11
-
-#ifdef _OPENMP
-            #pragma omp parallel for schedule(dynamic, 16)
-#endif
-
-            for (int y = 0; y < height; y++) { //data after ciecam
-                for (int x = 0; x < width; x++) {
-                    data[(y) * width + (x)] = lab->L[y][x];
-                    datanorm[(y) * width + (x)] = lab->L[y][x];
-
-                }
-            }
-
-            double nbs = 1.;
-            drd = std::max(drd, 1.); 
-            if (bl > 0.5f) {
-                nbs = (1.7 * (double) bl * drd);//take into account DR to increase variance in image source
-            }
-            if(!params->locallab.spots.at(sp).logcieq) {// not with log encoding Q
-                normalize_mean_dt(datanorm, datain, height * width, 1.f, 1.f, 0.f, 0.f, 0.f, 0.f, nbs);//normalize luminance
-            }
-
-#ifdef _OPENMP
-            #pragma omp parallel for schedule(dynamic, 16)
-#endif
-
-            for (int ir = 0; ir < height; ir++) {
-                for (int jr = 0; jr < width; jr++) {
-                    if(!params->locallab.spots.at(sp).logcieq) {// if not Log encoding ciecam
-                        data[ir * width + jr] = intp(bl, data[ir * width + jr], datanorm[ir * width + jr]);//blend with original
-                    } else {
-                        data[ir * width + jr] = intp(bl, data[ir * width + jr], datain[ir * width + jr]);//blend with original
-                    }
-                    lab->L[ir][jr] = data[ir * width + jr];
-                }
-            }
-        }
-
-        delete [] datain;
-        delete [] data;
-        delete [] datanorm;
-       
-        
-        
-    }
-
-
-
-}
-
-void ImProcFunctions::softproc(const LabImage* bufcolorig, const LabImage* bufcolfin, float rad, int bfh, int bfw, float epsilmax, float epsilmin, float thres, int sk, bool multiThread, int flag)
-{
-    if (rad != 0.f) {
-        array2D<float> ble(bfw, bfh);
-        array2D<float> guid(bfw, bfh);
-
-        if (flag == 0) {
-
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int ir = 0; ir < bfh; ir++) {
-                for (int jr = 0; jr < bfw; jr++) {
-                    guid[ir][jr] = Color::L2Y(bufcolorig->L[ir][jr]) / 32768.f;
-                    ble[ir][jr] = Color::L2Y(bufcolfin->L[ir][jr]) / 32768.f;
-                }
-            }
-
-            const float aepsil = (epsilmax - epsilmin) / 100.f;
-            const float bepsil = epsilmin; //epsilmax - 100.f * aepsil;
-            // const float epsil = aepsil * 0.1f * rad + bepsil;
-            const float epsil = aepsil * rad + bepsil;
-            const float blur = 10.f / sk * (thres + 0.f * rad);
-
-            rtengine::guidedFilter(guid, ble, ble, blur, epsil, multiThread, 4);
-
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int ir = 0; ir < bfh; ir++) {
-                for (int jr = 0; jr < bfw; jr++) {
-                    bufcolfin->L[ir][jr] =  Color::computeXYZ2LabY(32768.f * ble[ir][jr]);
-                }
-            }
-        } else if (flag == 1) {
-
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int ir = 0; ir < bfh; ir++)
-                for (int jr = 0; jr < bfw; jr++) {
-                    ble[ir][jr] = bufcolfin->L[ir][jr] / 32768.f;
-                    guid[ir][jr] = bufcolorig->L[ir][jr] / 32768.f;
-                }
-
-            const float aepsil = (epsilmax - epsilmin) / 1000.f;
-            const float bepsil = epsilmin; //epsilmax - 100.f * aepsil;
-            const float epsil = rad < 0.f ? 0.0001f : aepsil * rad + bepsil;
-            const float blur = rad < 0.f ? -1.f / rad : 1.f + rad;
-            const int r2 = rtengine::max(int(25 / sk * blur + 0.5f), 1);
-
-            rtengine::guidedFilter(guid, ble, ble, r2, epsil, multiThread);
-
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int ir = 0; ir < bfh; ir++) {
-                for (int jr = 0; jr < bfw; jr++) {
-                    bufcolfin->L[ir][jr] =  32768.f * ble[ir][jr];
-                }
-            }
-        } else if (flag == 2) {
-
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int ir = 0; ir < bfh; ir++)
-                for (int jr = 0; jr < bfw; jr++) {
-                    ble[ir][jr] = bufcolfin->L[ir][jr] / 32768.f;
-                    guid[ir][jr] = bufcolorig->L[ir][jr] / 32768.f;
-                }
-
-            const float aepsil = (epsilmax - epsilmin) / 1000.f;
-            const float bepsil = epsilmin; //epsilmax - 100.f * aepsil;
-            const float epsil = rad < 0.f ? 0.0001f : aepsil * 10.f * rad + bepsil;
-            //  const float epsil =  bepsil;
-            const float blur = rad < 0.f ? -1.f / rad : 0.00001f + rad;
-            const int r2 = rtengine::max(int(20.f / sk * blur + 0.000001f), 1);
-
-            rtengine::guidedFilter(guid, ble, ble, r2, epsil, multiThread);
-
-#ifdef _OPENMP
-            #pragma omp parallel for if(multiThread)
-#endif
-
-            for (int ir = 0; ir < bfh; ir++) {
-                for (int jr = 0; jr < bfw; jr++) {
-                    bufcolfin->L[ir][jr] =  32768.f * ble[ir][jr];
-                }
-            }
-        }
-    }
-}
-
-
-void ImProcFunctions::softprocess(const LabImage* bufcolorig, array2D<float> &buflight, float rad, int bfh, int bfw, double epsilmax, double epsilmin,  float thres, int sk, bool multiThread)
-{
-    float minlig = buflight[0][0];
-
-#ifdef _OPENMP
-    #pragma omp parallel for reduction(min:minlig) schedule(dynamic,16) if (multiThread)
-#endif
-
-    for (int ir = 0; ir < bfh; ir++) {
-        for (int jr = 0; jr < bfw; jr++) {
-            minlig = rtengine::min(buflight[ir][jr], minlig);
-        }
-    }
-
-    array2D<float> guidsoft(bfw, bfh);
-
-#ifdef _OPENMP
-    #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-    for (int ir = 0; ir < bfh; ir++) {
-        for (int jr = 0; jr < bfw; jr++) {
-            buflight[ir][jr] = LIM01((buflight[ir][jr] - minlig) / (100.f - minlig));
-            guidsoft[ir][jr] = bufcolorig->L[ir][jr] / 32768.f;
-        }
-    }
-
-    double aepsil = (epsilmax - epsilmin) / 90.0;
-    double bepsil = epsilmax - 100.0 * aepsil;
-    double epsil = aepsil * static_cast<double>(rad) + bepsil;
-    float blur = 1.f / sk * (thres + 0.8f * rad);
-    guidedFilter(guidsoft, buflight, buflight, blur, epsil,  multiThread, 4);
-
-
-#ifdef _OPENMP
-    #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-    for (int ir = 0; ir < bfh; ir++) {
-        for (int jr = 0; jr < bfw; jr++) {
-            buflight[ir][jr] = (100.f - minlig) * buflight[ir][jr] + minlig;
-        }
-    }
-}
-
-void ImProcFunctions::exlabLocal(local_params& lp, float strlap, int bfh, int bfw, int bfhr, int bfwr, LabImage* bufexporig, LabImage* lab, const LUTf& hltonecurve, const LUTf& shtonecurve, const LUTf& tonecurve, const float hueref, const float lumaref, const float chromaref)
-{
-    //BENCHFUN
-    //exposure local
-
-    constexpr float maxran = 65536.f;
-
-    if (lp.laplacexp == 0.f) {
-        lp.linear = 0.f;
-    }
-
-    const float linear = lp.linear;
-    int bw = bfw;
-    int bh = bfh;
-
-    if (linear > 0.f && lp.expcomp == 0.f) {
-        lp.expcomp = 0.001f;
-    }
-
-    const bool exec = (lp.expmet == 1 && linear > 0.f && lp.laplacexp > 0.1f);
-
-    if (!exec) { //for standard exposure
-
-        const float cexp_scale = std::pow(2.f, lp.expcomp);
-        const float ccomp = (rtengine::max(0.f, lp.expcomp) + 1.f) * lp.hlcomp / 100.f;
-        const float cshoulder = ((maxran / rtengine::max(1.0f, cexp_scale)) * (lp.hlcompthr / 200.f)) + 0.1f;
-        const float chlrange = maxran - cshoulder;
-        const float diffde = 100.f - lp.sensex;//the more scope, the less take into account dE for Laplace
-
-        if (!lp.invex) { // Laplacian not in inverse
-            bw = bfwr;
-            bh = bfhr;
-
-            //Laplacian PDE before exposure to smooth L, algorithm exposure leads to increase L differences
-            const std::unique_ptr<float[]> datain(new float[bfwr * bfhr]);
-            const std::unique_ptr<float[]> dataout(new float[bfwr * bfhr]);
-            const std::unique_ptr<float[]> dE(new float[bfwr * bfhr]);
-
-            deltaEforLaplace(dE.get(), diffde, bfwr, bfhr, bufexporig, hueref, chromaref, lumaref);
-
-            float alap = strlap * 600.f;
-            float blap = strlap * 100.f;
-            float aa = (alap - blap) / 50.f;
-            float bb = blap - 30.f * aa;
-
-            float lap;
-
-            if (diffde > 80.f) {
-                lap = alap;
-            } else if (diffde < 30.f) {
-                lap = blap;
-            } else {
-                lap = aa * diffde + bb;
-            }
-
-#ifdef _OPENMP
-            #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-            for (int y = 0; y < bfhr; y++) {
-                for (int x = 0; x < bfwr; x++) {
-                    datain[y * bfwr + x] = bufexporig->L[y][x];
-                }
-            }
-
-            MyMutex::MyLock lock(*fftwMutex);
-            ImProcFunctions::retinex_pde(datain.get(), dataout.get(), bfwr, bfhr, lap, 1.f, dE.get(), 0, 1, 1);//350 arbitrary value about 45% strength Laplacian
-#ifdef _OPENMP
-            #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-            for (int y = 0; y < bfhr; y++) {
-                for (int x = 0; x < bfwr; x++) {
-                    bufexporig->L[y][x] = dataout[y * bfwr + x];
-                }
-            }
-
-        }
-
-#ifdef _OPENMP
-        #pragma omp parallel for if (multiThread)
-#endif
-
-        for (int ir = 0; ir < bh; ir++) {//for standard with Laplacian in normal and without in inverse
-            for (int jr = 0; jr < bw; jr++) {
-                float L = bufexporig->L[ir][jr];
-                //highlight
-                const float hlfactor = (2 * L < MAXVALF ? hltonecurve[2 * L] : CurveFactory::hlcurve(cexp_scale, ccomp, chlrange, 2 * L));
-                L *= hlfactor;//approximation but pretty good with Laplacian and L < mean, hl aren't call
-                //shadow tone curve
-                L *= shtonecurve[2 * L];
-                //tonecurve
-                lab->L[ir][jr] = 0.5f * tonecurve[2 * L];
-            }
-        }
-    } else if (!lp.invex) { //for PDE algorithms
-        constexpr float kl = 1.f;
-        const float hlcompthr = lp.hlcompthr / 200.f;
-        const float hlcomp = lp.hlcomp / 100.f;
-
-#ifdef _OPENMP
-        #pragma omp parallel for if (multiThread)
-#endif
-
-        for (int ir = 0; ir < bfh; ir++) {
-            for (int jr = 0; jr < bfw; jr++) {
-                float L = bufexporig->L[ir][jr];
-                const float Llin = LIM01(L / 32768.f);
-                const float addcomp = linear * (-kl * Llin + kl);//maximum about 1 . IL
-                const float exp_scale = pow_F(2.f, lp.expcomp + addcomp);
-                const float shoulder = (maxran / rtengine::max(1.0f, exp_scale)) * hlcompthr + 0.1f;
-                const float comp = (rtengine::max(0.f, (lp.expcomp + addcomp)) + 1.f) * hlcomp;
-                const float hlrange = maxran - shoulder;
-
-                //highlight
-                const float hlfactor = (2 * L < MAXVALF ? hltonecurve[2 * L] : CurveFactory::hlcurve(exp_scale, comp, hlrange, 2 * L));
-                L *= hlfactor * pow_F(2.f, addcomp);//approximation but pretty good with Laplacian and L < mean, hl aren't call
-                //shadow tone curve
-                L *= shtonecurve[2 * L];
-                //tonecurve
-                lab->L[ir][jr] = 0.5f * tonecurve[2 * L];
-            }
-        }
-    }
-}
-
-void ImProcFunctions::addGaNoise(LabImage *lab, LabImage *dst, const float mean, const float variance, const int sk)
-{
-//   BENCHFUN
-//Box-Muller method.
-// add luma noise to image
-
-    srand(1);
-
-    const float variaFactor = SQR(variance) / sk;
-    constexpr float randFactor1 = 1.f / RAND_MAX;
-    constexpr float randFactor2 = (2.f * rtengine::RT_PI_F) / RAND_MAX;
-#ifdef _OPENMP
-    #pragma omp parallel if (multiThread)
-#endif
-    {
-        float z0, z1;
-        bool generate = false;
-#ifdef _OPENMP
-        #pragma omp for schedule(static) // static scheduling is important to avoid artefacts
-#endif
-
-        for (int y = 0; y < lab->H; y++) {
-            for (int x = 0; x < lab->W; x++) {
-                generate = !generate;
-                float kvar = 1.f;
-
-                if (lab->L[y][x] < 12000.f) {
-                    constexpr float ah = -0.5f / 12000.f;
-                    constexpr float bh = 1.5f;
-                    kvar = ah * lab->L[y][x] + bh;    //increase effect for low lights < 12000.f
-                } else if (lab->L[y][x] > 20000.f) {
-                    constexpr float ah = -0.5f / 12768.f;
-                    constexpr float bh = 1.f - 20000.f * ah;
-                    kvar = ah * lab->L[y][x] + bh;    //decrease effect for high lights > 20000.f
-                    kvar = kvar < 0.5f ? 0.5f : kvar;
-                }
-
-                float varia = SQR(kvar) * variaFactor;
-
-                if (!generate) {
-                    dst->L[y][x] = LIM(lab->L[y][x] + mean + varia * z1, 0.f, 32768.f);
-                    continue;
-                }
-
-                int u1 = 0;
-                int u2;
-
-                while (u1 == 0) {
-                    u1 = rand();
-                    u2 = rand();
-                }
-
-                float u1f = u1 * randFactor1;
-                float u2f = u2 * randFactor2;
-
-                float2 sincosval = xsincosf(2.f * rtengine::RT_PI_F * u2f);
-                float factor = std::sqrt(-2.f * xlogf(u1f));
-                z0 = factor * sincosval.y;
-                z1 = factor * sincosval.x;
-
-                dst->L[y][x] = LIM(lab->L[y][x] + mean + varia * z0, 0.f, 32768.f);
-
-            }
-        }
-    }
-}
+// addGaNoise moved to locallab/iplocallab_denoise.cc
 
 
 void ImProcFunctions::DeNoise_Local(int call, const struct local_params& lp, LabImage* originalmask, int levred, float hueref, float lumaref, float chromaref, LabImage* original, LabImage* transformed, const LabImage &tmp1, int cx, int cy, int sk)
@@ -9970,79 +6974,7 @@ void ImProcFunctions::transit_shapedetect2(int sp, float meantm, float stdtm, in
 
 
 
-void ImProcFunctions::exposure_pde(float * dataor, float * datain, float * dataout, int bfw, int bfh, float thresh, float mod)
-/* Jacques Desmis July 2019
-** adapted from Ipol Copyright 2009-2011 IPOL Image Processing On Line http://www.ipol.im/
-*/
-{
-
-    //BENCHFUN
-#ifdef RT_FFTW3F_OMP
-    if (multiThread) {
-        fftwf_init_threads();
-        fftwf_plan_with_nthreads(omp_get_max_threads());
-    }
-
-#endif
-    float *data_fft, *data_tmp, *data;
-
-    if (NULL == (data_tmp = (float *) fftwf_malloc(sizeof(float) * bfw * bfh))) {
-        fprintf(stderr, "allocation error\n");
-        abort();
-    }
-
-    ImProcFunctions::discrete_laplacian_threshold(data_tmp, datain, bfw, bfh, thresh);
-
-    if (NULL == (data_fft = (float *) fftwf_malloc(sizeof(float) * bfw * bfh))) {
-        fprintf(stderr, "allocation error\n");
-        abort();
-    }
-
-    if (NULL == (data = (float *) fftwf_malloc(sizeof(float) * bfw * bfh))) {
-        fprintf(stderr, "allocation error\n");
-        abort();
-    }
-
-    const auto dct_fw = fftwf_plan_r2r_2d(bfh, bfw, data_tmp, data_fft, FFTW_REDFT10, FFTW_REDFT10, FFTW_ESTIMATE | FFTW_DESTROY_INPUT);
-    fftwf_execute(dct_fw);
-
-    fftwf_free(data_tmp);
-
-    /* solve the Poisson PDE in Fourier space */
-    /* 1. / (float) (bfw * bfh)) is the DCT normalisation term, see libfftw */
-    ImProcFunctions::rex_poisson_dct(data_fft, bfw, bfh, 1. / (double)(bfw * bfh));
-
-    const auto dct_bw = fftwf_plan_r2r_2d(bfh, bfw, data_fft, data, FFTW_REDFT01, FFTW_REDFT01, FFTW_ESTIMATE | FFTW_DESTROY_INPUT);
-    fftwf_execute(dct_bw);
-    fftwf_destroy_plan(dct_fw);
-    fftwf_destroy_plan(dct_bw);
-    fftwf_free(data_fft);
-    fftwf_cleanup();
-
-#ifdef RT_FFTW3F_OMP
-
-    if (multiThread) {
-        fftwf_cleanup_threads();
-    }
-
-#endif
-
-    normalize_mean_dt(data, dataor, bfw * bfh, mod, 1.f, 0.f, 0.f, 0.f, 0.f, 1.);
-    {
-
-#ifdef _OPENMP
-        #pragma omp parallel for
-#endif
-
-        for (int y = 0; y < bfh ; y++) {
-            for (int x = 0; x < bfw; x++) {
-                dataout[y * bfw + x] = clipLoc(data[y * bfw + x]);
-            }
-        }
-    }
-
-    fftwf_free(data);
-}
+// exposure_pde moved to locallab/iplocallab_exposure.cc
 
 void ImProcFunctions::fftw_convol_blur(float * input, float * output, int bfw, int bfh, float radius, int fftkern, int algo)
 {
@@ -19461,86 +16393,8 @@ void ImProcFunctions::Lab_Local(
         }
     }
 
-// soft light and retinex_pde
-    if ((lp.strng > 1.f || lp.prevdE) && call <= 3 && lp.sfena) {
-        int ystart = rtengine::max(static_cast<int>(lp.yc - lp.lyT) - cy, 0);
-        int yend = rtengine::min(static_cast<int>(lp.yc + lp.ly) - cy, original->H);
-        int xstart = rtengine::max(static_cast<int>(lp.xc - lp.lxL) - cx, 0);
-        int xend = rtengine::min(static_cast<int>(lp.xc + lp.lx) - cx, original->W);
-        int bfh = yend - ystart;
-        int bfw = xend - xstart;
-        //variable for fast FFTW
-        int bfhr = bfh;
-        int bfwr = bfw;
-
-        if (bfw >= mSP && bfh >= mSP) {
-
-            if (lp.softmet == 1) {
-                optfft(N_fftwsize, bfh, bfw, bfhr, bfwr, lp, original->H, original->W, xstart, ystart, xend, yend, cx, cy, lp.fullim);
-            }
-
-            const std::unique_ptr<LabImage> bufexporig(new LabImage(bfw, bfh));
-            const std::unique_ptr<LabImage> bufexpfin(new LabImage(bfw, bfh));
-
-#ifdef _OPENMP
-            #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-            for (int y = ystart; y < yend; y++) {
-                for (int x = xstart; x < xend; x++) {
-                    bufexporig->L[y - ystart][x - xstart] = original->L[y][x];
-                    bufexporig->a[y - ystart][x - xstart] = original->a[y][x];
-                    bufexporig->b[y - ystart][x - xstart] = original->b[y][x];
-                }
-            }
-
-            bufexpfin->CopyFrom(bufexporig.get(), multiThread);
-            SoftLightParams softLightParams;
-            softLightParams.enabled = true;
-            softLightParams.strength = lp.strng;
-
-            if (lp.softmet == 0) {
-                ImProcFunctions::softLight(bufexpfin.get(), softLightParams);
-            } else if (lp.softmet == 1) {
-                const std::unique_ptr<float[]> datain(new float[bfwr * bfhr]);
-                const std::unique_ptr<float[]> dataout(new float[bfwr * bfhr]);
-                const std::unique_ptr<float[]> dE(new float[bfwr * bfhr]);
-
-                deltaEforLaplace(dE.get(), lp.lap, bfwr, bfhr, bufexpfin.get(), hueref, chromaref, lumaref);
-
-#ifdef _OPENMP
-                #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-                for (int y = 0; y < bfhr; y++) {
-                    for (int x = 0; x < bfwr; x++) {
-                        datain[y * bfwr + x] = bufexpfin->L[y][x];
-                    }
-                }
-
-                const int showorig = lp.showmasksoftmet >= 5 ? 0 : lp.showmasksoftmet;
-                MyMutex::MyLock lock(*fftwMutex);
-                ImProcFunctions::retinex_pde(datain.get(), dataout.get(), bfwr, bfhr, 8.f * lp.strng, 1.f, dE.get(), showorig, 1, 1);
-#ifdef _OPENMP
-                #pragma omp parallel for schedule(dynamic,16) if (multiThread)
-#endif
-
-                for (int y = 0; y < bfhr; y++) {
-                    for (int x = 0; x < bfwr; x++) {
-                        bufexpfin->L[y][x] = dataout[y * bfwr + x];
-                    }
-                }
-            }
-
-            transit_shapedetect2(sp, 0.f, 0.f, call, 3, bufexporig.get(), bufexpfin.get(), nullptr, hueref, chromaref, lumaref, sobelref, 0.f, nullptr, lp, original, transformed, nullptr, LocalLabGradientMode::STANDARD, cx, cy, sk);
-
-            if (lp.recur) {
-                original->CopyFrom(transformed, multiThread);
-                float avge;
-                calc_ref(sp, original, transformed, 0, 0, original->W, original->H, sk, huerefblur, chromarefblur, lumarefblur, hueref, chromaref, lumaref, sobelref, avge, locwavCurveden, locwavdenutili);
-            }
-        }
-    }
+// Soft Light tool - implementation in locallab/iplocallab_softlight.inc
+#include "locallab/iplocallab_softlight.inc"
 
     //local contrast
     bool wavcurve = false;
