+++ rtengine/locallab/iplocallab_common.cc
@@ -1,0 +1,32 @@
+/*
+ *  This file is part of RawTherapee.
+ *
+ *  Copyright (c) 2004-2010 Gabor Horvath <hgabor@rawtherapee.com>
+ *  Copyright (c) 2016-2024 Jacques Desmis <jdesmis@gmail.com>
+ *
+ *  RawTherapee is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  RawTherapee is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with RawTherapee.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+// This file contains common utilities for locallab tools.
+// Currently a stub - implementations will be moved here incrementally.
+
+#include "iplocallab.h"
+
+namespace rtengine {
+
+// Placeholder - common functions will be moved here from iplocallab.cc
+// The actual implementations remain in iplocallab.cc for now to ensure
+// the refactoring can be done incrementally without breaking the build.
+
+} // namespace rtengine
